% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-external.R
\name{addExpressionData}
\alias{addExpressionData}
\title{Include custom normalized data in the DNEAresults object}
\usage{
addExpressionData(object, data)
}
\arguments{
\item{object}{A \code{DNEAresults} object}

\item{data}{An \emph{m x n} numeric matrix of custom-normalized expression expression data}
}
\value{
A \code{DNEAresults} object with the added expression data in the @assays slot
}
\description{
This function allows the user to input custom-normalized data into the DNEAresults object for use in DNEA analysis.
}
\examples{
#import example data
data(TEDDY)
data(dnw)
data(T1Dmeta)

#transpose TEDDY data
TEDDY <- t(TEDDY)

#make sure metadata and expression data are in same order
T1Dmeta <- T1Dmeta[rownames(TEDDY),]

dat <- list('DM:control' = TEDDY[T1Dmeta$group == "DM:control",],
            'DM:case' = TEDDY[T1Dmeta$group == "DM:case",])

#log-transform and median center the expression data without scaling
newdat <- NULL
for(cond in dat){
 for(i in seq(1, ncol(cond))){
  my_median = median(cond[, i], na.rm = TRUE)
   my_range = range(cond[, i], na.rm = TRUE)
    scale_factor = max(abs(my_range-my_median))
    cond[, i] <- (cond[, i] - my_median) / scale_factor
  }
  newdat <- rbind(newdat, cond)
}

#reorder to match dnw
newdat <- newdat[sampleNames(dnw), featureNames(dnw)]
newdat <- t(newdat)

#add data
dnw <- addExpressionData(object = dnw, data = newdat)

}
\seealso{
\code{\link{createDNEAobject}}, \code{DNEAresults},
}
\author{
Christopher Patsalis
}
