% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{adjacencyGraph}
\alias{adjacencyGraph}
\alias{adjacencyGraph<-}
\alias{adjacencyGraph,DNEAresults-method}
\alias{adjacencyGraph<-,DNEAresults-method}
\alias{adjacencyGraph,consensusClusteringResults-method}
\alias{adjacencyGraph<-,consensusClusteringResults-method}
\title{adjacencyGraph retrieves the adjacency graph for the case, control, or joint network}
\usage{
adjacencyGraph(x, graph)

adjacencyGraph(x, graph) <- value

\S4method{adjacencyGraph}{DNEAresults}(x, graph)

\S4method{adjacencyGraph}{DNEAresults}(x, graph) <- value

\S4method{adjacencyGraph}{consensusClusteringResults}(x, graph)

\S4method{adjacencyGraph}{consensusClusteringResults}(x, graph) <- value
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}

\item{graph}{the adjacency graph to retrieve. Values can be a provided condition name for either independent network,
or "joint_graph" for the total network graph}
}
\value{
The input object after filling the @adjacency_matrix slot or a graph corresponding to the specified
adjacency graph
}
\description{
The function takes as input a pcorNetwork, DNEAobject, DNEAobject_collapsed, or consensusClusteringResults object
and returns the adjacency graph made for the case, control, or joint network determined via runConsensusCluster().
\strong{FOR INTERNAL USE ONLY - The graph can also be inserted into the object via the setter functionality of the function.}
}
\keyword{internal}
