% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{adjacencyMatrix}
\alias{adjacencyMatrix}
\alias{adjacencyMatrix<-}
\alias{adjacencyMatrix,DNEAresults-method}
\alias{adjacencyMatrix<-,DNEAresults-method}
\title{adjacencyMatrix retrieves the weighted or unweighted adjacency matrix}
\usage{
adjacencyMatrix(x, weighted)

adjacencyMatrix(x, weighted) <- value

\S4method{adjacencyMatrix}{DNEAresults}(x, weighted = FALSE)

\S4method{adjacencyMatrix}{DNEAresults}(x, weighted = FALSE) <- value
}
\arguments{
\item{a}{pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
The input object after filling the @adjacency_matrix slot or a matrix corresponding to the adjacency matrix
}
\description{
The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object
and returns the weighted or unweighted adjacency matrix determined via getNetworks().
\strong{FOR INTERNAL USE ONLY - The matrix can also be inserted into the object via the setter functionality of the function.}
}
\keyword{internal}
