% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEM.R, R/tuning.R
\name{adjDGlasso_minimal}
\alias{adjDGlasso_minimal}
\title{adjDGlasso_minimal constructs the debiased glasso model}
\usage{
adjDGlasso_minimal(
  X,
  weights = 1,
  theta_star = NULL,
  lambda = NULL,
  quiet = FALSE,
  zero.edge = NULL
)

adjDGlasso_minimal(
  X,
  weights = 1,
  theta_star = NULL,
  lambda = NULL,
  quiet = FALSE,
  zero.edge = NULL
)
}
\arguments{
\item{X}{A matrix of expression data wherein the samples are rows and features are columns.}

\item{weights}{A matrix of selection weights determined via stability selection to be integrated
into the model. Default is 1.}

\item{theta_star}{The true precision matrix. Default is NULL}

\item{lambda}{The optimized lambda hyper-parameter. The default is NULL}

\item{quiet}{A boolean that indicates whether progress messages should be displayed. The default
value is FALSE}

\item{zero.edge}{Indices of entries of inverse covariance to be constrained to zero (to be passed
to glasso). The default is NULL}
}
\value{
An adjacency matrix for the data network estimated by the glasso model

An adjacency matrix for the data network estimated by the glasso model
}
\description{
This function takes the expression data and optimized lambda as input and fits the glasso model

This function takes the expression data and optimized lambda as input and fits the glasso model
}
