% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgsa_complex.R
\name{approxVarEst}
\alias{approxVarEst}
\title{approxVarEst does something}
\usage{
approxVarEst(se0, sg0, D, r, n_vec, control = NULL)
}
\value{
what it returns
}
\description{
more info here
}
\keyword{internal}
