% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{BICscores}
\alias{BICscores}
\alias{BICscores<-}
\alias{BICscores,DNEAobj-method}
\alias{BICscores<-,DNEAobj-method}
\title{Access the BIC scores for each lambda value evaluated}
\usage{
BICscores(x)

BICscores(x) <- value

\S4method{BICscores}{DNEAobj}(x)

\S4method{BICscores}{DNEAobj}(x) <- value
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}

\item{value}{a list containing lists that consist of the liklihood and
BIC score for tested lambda values}
}
\value{
The optimized lambda hyperparameter.
}
\description{
The function takes as input a \code{\link{DNEAobj}} object and returns
the BIC values for each lambda tested during hyperparameter optimization
performed via BICtune().
}
\examples{
#import example data
data(dnw)

BICscores(dnw)
}
\seealso{
\code{\link{BICtune}}
}
\author{
Christopher Patsalis
}
