% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{BICscores.pcorNetwork}
\alias{BICscores.pcorNetwork}
\title{BICscores returns the BIC scores for each lambda value evaluated}
\usage{
BICscores.pcorNetwork(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
The optimized lambda hyperparameter
}
\description{
BICscores returns the BIC scores for each lambda value evaluated
}
\keyword{internal}
