% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{BICtune}
\alias{BICtune}
\title{BICtune Optimizes the Lambda parameter for glasso}
\usage{
BICtune(
  object,
  lambda_values,
  eps_threshold = 1e-06,
  eta_value = 0.1,
  BPPARAM = bpparam()
)
}
\arguments{
\item{object}{A DNEA object}

\item{lambda_values}{An optional list of lambda values to fit a model and calculate the BIC score.
If not provided, a set of lambda values are chosen based on the size of the dataset.}

\item{eps_threshold}{A significance cut-off for thresholding network edges
The default value is 1e-06.}

\item{eta_value}{default parameter ??. Default is 0.1}

\item{BPPARAM}{A BiocParallel object}
}
\value{
A DNEAobject containing the BIC and liklihood scores for every lambda value tested, as well as
the optimized lambda value
}
\description{
This function will calculate the Bayesian information criterion (BIC) and liklihood for a range of lambda values
determined by the number of features within the dataset. The lambda value with the lowest (BIC) score is
chosen for analysis. It takes a DNEAobject as input and has an option that allows the function to be run
in parallel.
}
