% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgsa_complex.R
\name{call.netgsa}
\alias{call.netgsa}
\title{call.netgsa does this}
\usage{
call.netgsa(D, x, y, B, varEstCntrl)
}
\arguments{
\item{D}{stuff}

\item{x}{stuff}

\item{y}{stuff}

\item{B}{stuff}

\item{varEstCntrl}{stuff}
}
\value{
what it returns
}
\description{
more info here
}
