% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEMtuning.R
\name{CGM_AHP_stabsel_subsample}
\alias{CGM_AHP_stabsel_subsample}
\title{CGM_AHP_stabsel_subsample performs stability selection WITH additional subsampling}
\usage{
CGM_AHP_stabsel_subsample(
  listX,
  X,
  init_param,
  lastar,
  seed.base = 100,
  eta = 0.01,
  limkappa = 1e+06,
  main.seed = 101
)
}
\arguments{
\item{listX}{A list containing matrices of the expression data split by condition. The samples
are in rows and the features are columns.}

\item{X}{A vector corresponding to the number of reps to be run in stability selection}

\item{lastar}{the optimized lambda parameter}

\item{eta}{default parameter ??. Default is 0.01}

\item{limkappa}{default parameter that acts as the limit for the condition number of the sample cov.
Default is 1e+6}

\item{main.seed}{Sets the seed for random number generation. This ensures reproducibility.}
}
\value{
a precision matrix for the network corresponding to the input data created via that rep
}
\description{
This function will take as input the expression data and optimized lambda in order to randomly sample
the data to perform stability selection. The function is based on the method described by Guo et al.
(2011) and is designed to be run with some variation of the lapply family of functions. A key
difference between this function and the related CGM_AHP_stabsel is that random sampling is done
in a way that evens out sample distribution across the two conditions
}
