% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEM.R
\name{CGM_AHP_train}
\alias{CGM_AHP_train}
\title{CGM_AHP_train will train a glasso model}
\usage{
CGM_AHP_train(
  trainX,
  trainY,
  lambda_value,
  adaptive_weight = array(1, c(length(unique(trainY)), ncol(trainX), ncol(trainX))),
  eta = 0.01,
  limkappa = 1e+06
)
}
\arguments{
\item{trainX}{a matrix of expression data wherein the samples are rows and features are columns.}

\item{trainY}{a vector of corresponding conditions for samples in trainX}

\item{lambda_value}{the lambda hyperparameter to use in the glasso model}

\item{adaptive_weight}{default parameter used to calculate the penalty matrix}

\item{eta}{default parameter ??. Default is 0.01}

\item{limkappa}{default parameter that acts as the limit for the condition number of the sample cov.
Default is 1e+6}
}
\value{
The ouput of a glasso model
}
\description{
This function takes the data and corresponding condition, as well as a lambda value as input and
trains a glasso model, outputting the results. This code comes from Guo et al. (2011)
}
\author{
Jing Ma
}
\keyword{internal}
