% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEM.R, R/tuning.R
\name{CGM_AHP_tune}
\alias{CGM_AHP_tune}
\title{CGM_AHP_tune optimizes the tuning parameter for Guo et al. (2011) method in CGM_AHP_train}
\usage{
CGM_AHP_tune(
  trainX,
  testX,
  model,
  X,
  BIC = FALSE,
  eps = 1e-06,
  eta = 0.01,
  limkappa = 1e+06
)

CGM_AHP_tune(
  trainX,
  testX,
  model,
  X,
  BIC = FALSE,
  eps = 1e-06,
  eta = 0.01,
  limkappa = 1e+06
)
}
\arguments{
\item{trainX}{a matrix of expression data wherein the samples are rows and features are columns.}

\item{testX}{a matrix of expression data wherein the samples are rows and features are columns.
testX should be identical to trainX.}

\item{model}{a vector of corresponding conditions for samples in trainX}

\item{X}{a vector of lambda values tested to find the optimal hyper-parameter}

\item{BIC}{a boolean indicating whether or not to calculate the BIC score for each lambda.
Default is FALSE.}

\item{eps}{A significance cut-off for thresholding network interactions.
Default is 1e-06}

\item{eta}{default parameter ??. Default is 0.01}

\item{limkappa}{default parameter that acts as the limit for the condition number of the sample cov.
Default is 1e+6}
}
\value{
A list containing the BIC and liklihood score for each lambda parameter evaluated.

A list containing the BIC and liklihood score for each lambda parameter evaluated.
}
\description{
This function takes the data and corresponding condition, as well as a list of lambda values
as input in order to optimize the lambda hyper-parameter.

This function takes the data and corresponding condition, as well as a list of lambda values
as input in order to optimize the lambda hyper-parameter.
}
