% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanFeatureNames}
\alias{cleanFeatureNames}
\title{cleanFeatureNames retrieves the feature names from clean_feature names in the metadata slot.}
\usage{
cleanFeatureNames(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
A vector of modified feature names
}
\description{
This function takes in a DNEAobject and return the feature names modified to avoid errors with
native R clashes. The names are located in the metadata slot of the object.
}
