% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{collapsed_DNEAresults-class}
\alias{collapsed_DNEAresults-class}
\title{collapse_DNEAresults}
\value{
A collapsed_DNEAresults object
}
\description{
An s4 class to represent the DNEA workflow, including collapsing features. This class inherits everything from
\code{\link{DNEAresults-class}}
}
\section{Slots}{

\describe{
\item{\code{original_experiment}}{The DNEAresults object input to \code{\link{aggregateFeatures}}}

\item{\code{feature_membership}}{A data.frame containing all of the features from the original input data and their
corresponding group membership in the collapsed data.}
}}

\seealso{
\code{\link{aggregateFeatures}}, \code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
