% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseNodes-fun.R
\name{collapseNodes_cor}
\alias{collapseNodes_cor}
\title{collapseNodes_cor will collapse nodes based on correlation}
\usage{
collapseNodes_cor(dat, correlation_threshold = 0.9)
}
\arguments{
\item{dat}{unscaled expression data}

\item{correlation_threshold}{correlation threshold}
}
\value{
collapsed data
}
\description{
more info about correlation
}
\keyword{internal}
