% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseNodes-fun.R
\name{collapseNodes_hybrid}
\alias{collapseNodes_hybrid}
\title{collapseNodes_hybrid node-collapsing based on correlations and user-supplied metabolite groups}
\usage{
collapseNodes_hybrid(dat, metabolite_groups, correlation_threshold = 0.9)
}
\arguments{
\item{dat}{unscaled expression data}

\item{metabolite_groups}{metabolite groupings}

\item{correlation_threshold}{correlation threshold}
}
\value{
collapsed data
}
\description{
more info about correlation
}
\keyword{internal}
