% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseNodes-fun.R
\name{collapseNodes_knowledge}
\alias{collapseNodes_knowledge}
\title{collapseNodes_knowledge node-collapsing based on user-supplied metabolite groups}
\usage{
collapseNodes_knowledge(dat, metabolite_groups)
}
\arguments{
\item{dat}{unscaled expression data}

\item{metabolite_groups}{metabolite groupings}
}
\value{
collapsed data
}
\description{
more info about correlation
}
\keyword{internal}
