% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{condition}
\alias{condition}
\title{setter function for NormalExpression data
condition retrieves the condition values for each sample from the metadata slot.}
\usage{
condition(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
A vector of the condition values
}
\description{
This function takes in a DNEAobject and return the condition values located in the
metatdata slot of the object. You can also use it to input conditions data into the DNEAobject
}
