% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conditionLevels.pcorNetwork}
\alias{conditionLevels.pcorNetwork}
\title{conditionLevels retrieves the unique condition names for the dataset}
\usage{
conditionLevels.pcorNetwork(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
A vector of the unique condition labels
}
\description{
This function takes in a DNEAobject and return the unique condition labels located in the
dataset_summary slot of the object
}
