% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{conditionLevels}
\alias{conditionLevels}
\alias{conditionLevels,DNEAobject-method}
\title{conditionLevels returns the group information for samples in a DNEAobject,
or DNEAobject_collapsed}
\usage{
conditionLevels(x)

\S4method{conditionLevels}{DNEAobject}(x)
}
\arguments{
\item{x}{A DNEAobject or DNEAobject_collapsed}
}
\value{
A vector of the unique condition labels
}
\description{
This function takes in a DNEAobject or DNEAobject_collapsed and return the unique condition
labels located in the dataset_summary slot of the object
}
