% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{conditions}
\alias{conditions}
\alias{conditions<-}
\alias{conditions,DNEAobject-method}
\alias{conditions<-,DNEAobject-method}
\title{condition retrieves the condition values for each sample from the metadata slot.}
\usage{
conditions(x)

conditions(x) <- value

\S4method{conditions}{DNEAobject}(x)

\S4method{conditions}{DNEAobject}(x) <- value
}
\arguments{
\item{x}{A DNEAobject, or DNEAobject_collapsed object}
}
\value{
A vector of the condition values
}
\description{
This function takes in a DNEAobject, or DNEAobject_collapsed object and
return the condition values located in the metatdata slot of the object
}
