% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{consensusClusteringResults-class}
\alias{consensusClusteringResults-class}
\alias{consensusClusteringResults}
\title{consensusClusteringResults}
\value{
A consensusClusteringResults object
}
\description{
An s4 class to represent the results from consensus
clustering within DNEA
}
\section{Slots}{

\describe{
\item{\code{summary}}{a data frame containing the sub networks
as rows and summary information as columns. The columns
include: number_of_nodes, number_of_edges,
number_of_DE_nodes, and number_of_DE_edges.}

\item{\code{subnetwork_membership}}{A data frame with the same
number of rows as features in the data, and a column
indicating which sub network a given feature belongs
to, if any.}

\item{\code{adjacency_graph}}{The resulting adjacency graph
from \code{\link{igraph}} created after
consensus clustering.}
}}

\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
