% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDNEAobject.R
\name{createDNEAobject}
\alias{createDNEAobject}
\title{Initializs the DNEAobject}
\usage{
createDNEAobject(
  project_name,
  expression_data,
  scaled_expression_data,
  control,
  case
)
}
\arguments{
\item{project_name}{A string containing an identifying name for the analysis}

\item{expression_data}{A matrix or dataframe of un-scaled expression data. The sample names should be rownames
and the feature names should be column names. Column 1 should be a factor of the two conditions, followed by
the numeric expression data}

\item{scaled_expression_data}{A matrix or dataframe similar to expression_data but for scaled data}

\item{control}{A string corresponding to the name of condition 1 in column one of the data matrix}

\item{case}{A string corresponding to the name of condition 2 in column one of the data matrix}
}
\value{
DNEAobject containing the scaled and un-scaled (if provided) data in assays, as well as sample number,
feature number, sample names, feature names, and condition number in metadata
}
\description{
Takes the input data and restructures it to create the DNEAobject and expression_data is scaled if
scaled_exprssion_data is not provided. Diagnostics are performed on the scaled data and the minimum
eigen value and condition number are calculated for the whole dataset and each condition, specified by the
control and case inputs, separated.The output will inform the user if Feature Reduction should be performed prior
to continuing the analysis.
}
