% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{dataDiagnostics}
\alias{dataDiagnostics}
\title{Datadiagnostics will run Differential Expression on Nodes and check data stability prior to analysis}
\usage{
dataDiagnostics(mat, condition_values, conditions)
}
\arguments{
\item{mat}{A matrix of expression data}

\item{condition_values}{A vector of condition names present in the dataset}

\item{conditions}{A vector of condition labels corresponding to the samples in mat}
}
\value{
A DNEA object containing an initialized node_list. Differential Expression is performed on
the features if un-scaled data is provided. The min eigen value and condition number is also
printed for the whole dataset as well as each condition.
}
\description{
This function first takes in a DNEAobject and runs diagnostics on the scaled expression data in the
scaled_expression_data value of the assays tab. It will then calculate the minimum eigen value and
condition number for the whole dataset, as well as for each condition and inform the user if they can
proceed with the analysis or collapse the features using the reduceFeatures() function prior.
}
\keyword{internal}
