% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{datasetSummary}
\alias{datasetSummary}
\alias{datasetSummary,DNEAobj-method}
\title{Access the dataset_summary slot of a DNEAobj object}
\usage{
datasetSummary(x)

\S4method{datasetSummary}{DNEAobj}(x)
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}
}
\value{
The numbers of samples/features and diagnostic values of the
input data calculated by \code{\link{createDNEAobject}}.
}
\description{
This function prints to console the number of samples, number of
features, and diagnostic values of the input data to
\code{\link{createDNEAobject}} at initiation of the DNEA workflow.
}
\examples{
#import example data
data(dnw)

datasetSummary(dnw)
}
\seealso{
\code{\link{createDNEAobject}},\code{\link{aggregateFeatures}}
}
\author{
Christopher Patsalis
}
