% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics<-}
\alias{diagnostics,DNEAresults-method}
\alias{diagnostics<-,DNEAresults-method}
\alias{diagnostics,DNEAinputSummary-method}
\alias{diagnostics<-,DNEAinputSummary-method}
\title{Retrieve the diagnostic values for the input expression data}
\usage{
diagnostics(x)

diagnostics(x) <- value

\S4method{diagnostics}{DNEAresults}(x)

\S4method{diagnostics}{DNEAresults}(x) <- value

\S4method{diagnostics}{DNEAinputSummary}(x)

\S4method{diagnostics}{DNEAinputSummary}(x) <- value
}
\arguments{
\item{a}{\code{DNEAresults} object or \code{DNEAinputSummary} object}
}
\value{
Returns the diagnostic values for the input expression data.
}
\description{
This function retrieves teh diagnostic values calculated for the input expression data to \code{\link{createDNEAobject}} \cr
}
\examples{
#import example data
data(dnw)

diagnostics(dnw)
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
