% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-here.R
\docType{data}
\name{dnw}
\alias{dnw}
\title{Example results for DNEA}
\format{
A DNEA results object after completing a DNEA experiment.
}
\source{
The data the results of the full DNEA workflow performed using
the \code{\link{TEDDY}} example data, as described above.
}
\usage{
data("dnw")
}
\value{
A \code{\link{DNEA}} object containing the results of a
DNEA experiment.
}
\description{
"dnw" is a DNEA object containing the results for the full
DNEA workflow on the \code{\link{TEDDY}} example data. The
exact workflow to produce these results can be replicated by
following the package vignette accessed by entering
browseVignettes("DNEA") in the console. 1000 replicates were
performed during stability selection \emph{with} the
subsampling protocol. The lambda value used during joint
estimation was aproximated as
\deqn{\lambda = \sqrt{ \ln (num. features) / num. samples}}{ lambda = sqrt(ln(num. features) / num. samples)}
}
\keyword{datasets}
