% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgsa_complex.R
\name{edgelist2adj}
\alias{edgelist2adj}
\title{edgelist2adj does this}
\usage{
edgelist2adj(file, vertex.names, mode = c("directed", "undirected"))
}
\value{
returns this
}
\description{
more info
}
\keyword{internal}
