% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{expressionData}
\alias{expressionData}
\alias{expressionData,DNEAobj-method}
\title{Access expression data within a DNEAobj object,}
\usage{
expressionData(x, assay)

\S4method{expressionData}{DNEAobj}(
  x,
  assay = c("input_data", "log_input_data", "scaled_expression_data")
)
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}

\item{assay}{A character string corresponding to the data to retrieve:
"input_data" retrieves the data as it was input, "log_input_data"
retrieves the input_data after log transforming, and
"scaled_expression_data" retrieves the log-scaled data matrices.}
}
\value{
The expression matrix specified by the user.
}
\description{
This function accesses the expression data stored in the @assays
slot of the \code{\link{DNEAobj}} object. The output is an \emph{n x m}
matrix with one row for each sample and one column for each feature
in the data.
}
\examples{
#import example data
data(dnw)

expressionData(x=dnw, assay="input_data")
}
\seealso{
\code{\link{createDNEAobject}},\code{\link{aggregateFeatures}}
}
\author{
Christopher Patsalis
}
