% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expressionData.pcorNetwork}
\alias{expressionData.pcorNetwork}
\title{expressionData retrieves un-scaled expression data from the assays slot.}
\usage{
expressionData.pcorNetwork(x, type = c("input", "normalized"))
}
\arguments{
\item{x}{A DNEAobject}

\item{type}{"input" will return the original data, and "normalized" will return the scaled data}
}
\value{
The indicated expression expression matrix
}
\description{
This function takes in a DNEAobject and return the un-scaled expression matrix located in the
assays slot of the object. You can also use it to input expression data into the DNEAobject
}
