% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expressionData}
\alias{expressionData}
\title{expressionData retrieves un-scaled expression data from the assays slot.}
\usage{
expressionData(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
the expression matrix
}
\description{
This function takes in a DNEAobject and return the un-scaled expression matrix located in the
assays slot of the object. You can also use it to input expression data into the DNEAobject
}
