% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{featureNames.pcorNetwork}
\alias{featureNames.pcorNetwork}
\title{featureNames retrieves the feature names from the metadata slot.}
\usage{
featureNames.pcorNetwork(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
A vector of feature names
}
\description{
This function takes in a DNEAobject and return the feature names located in the
metadata slot of the object.
}
