% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{featureNames}
\alias{featureNames}
\title{setter function for Samples}
\usage{
featureNames(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
A vector of feature names
}
\description{
featureNames retrieves the feature names from the metadata slot.
}
\details{
This function takes in a DNEAobject and return the feature names located in the
metadata slot of the object.
}
