% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filterNetworks}
\alias{filterNetworks}
\title{filterNetworks() will reduce the adjacency matrices to only the edges that meet the filter conditions}
\usage{
filterNetworks(object, pcor, top_percent_edges)
}
\arguments{
\item{object}{A DNEAresults object}

\item{pcor}{A pcor value of which to threshold the adjacency matrices. Edges with pcor values <= to this value will be removed.}

\item{top_percent_edges}{A value between 0-1 that corresponds to the top x\% edges to keep in the networks
ie. top_percent_edges = 0.1 will keep only the top 10\% strongest edges in the networks.}

\item{...}{Additional arguments to be passed to filterNetworks(). See \strong{Details} for more information.}
}
\value{
The input object after filtering the egdes in the network according to the specified parameters
}
\description{
filterNetworks() takes as input a DNEAresults object and allows the user to filter the network edges by one of two ways:\n
\enumerate{
\item The networks can be filtered to only include edges greater than or equal to a specified partial correlation (pcor) value.\n
\item The networks can be filtered to only include the strongest x\% of edges by pcor value.\n
Filtering is performed on the case and control adjacency matrix separately.
}
}
