% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_lib.R
\name{getConsensusMatrix}
\alias{getConsensusMatrix}
\title{getConsensusMatrix will calculate the consensus matrix of the network}
\usage{
getConsensusMatrix(cl)
}
\arguments{
\item{cl}{the results from the consensus clustering algorithm}
}
\value{
The corresponding consensus matrix
}
\description{
The function takes as input the results from the consensus clustering algorithm and calculates
the consensus matrix
}
\keyword{internal}
