% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getNetworkFiles}
\alias{getNetworkFiles}
\title{getNetworkFiles will save the node and edge information}
\usage{
getNetworkFiles(object, file_path)
}
\arguments{
\item{object}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}

\item{file_path}{The filepath to save the node and edge lists to. If \strong{NULL}, the files will be saved to the working
directory}
}
\value{
The same object as input and saves the node and edge information as .csv files in
the working directory
}
\description{
This function will save the node and edge information as .csv files in the working directory.
The files are formatted for input into Cytoscape.
}
