% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{getNeworks}
\alias{getNeworks}
\title{Creates the network model using glasso}
\usage{
getNeworks(object, eps = 1e-06)
}
\arguments{
\item{object}{A DNEAobject}

\item{eps}{A significance cut-off for thresholding the adjacency matrix. This can be left as default}
}
\value{
A DNEA object containing an adjacency matrix for the data network,
determined by the glasso model
}
\description{
This function takes in a DNEAobjct and fits a glasso model using the optimized lambda value
determined via BICtune() or otherwise specified. If selection probabilites for each node were
calculated using stabilitySelection(), those are also utilized. An adjacency matrix and a
thresholded adjacency matrix, determiend by the eps value, is output.
}
