% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lambdas2Test}
\alias{lambdas2Test}
\title{lambdas2Test returns the lambda value used in analysis.}
\usage{
lambdas2Test(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
The lambda values to evaluate in optimization
}
\description{
lambdas2Test returns the lambda value used in analysis.
}
