% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabDE.R
\name{metabDE}
\alias{metabDE}
\title{Performs differential expression on features of input data}
\usage{
metabDE(mat, condition_values, conditions)
}
\arguments{
\item{mat}{an mxn expression matrix with features as columns}

\item{condition_values}{a vector of the two possible condition values}

\item{conditions}{a vector containing the condition labels corresponding to the samples in mat}
}
\value{
A dataframe containing the differential expression results
}
\description{
Performs differential expression on features of the input data provided using student's T-Test and multiple hypothesis
correction via Benjamini-Hochberg
}
\keyword{internal}
