% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgsa_complex.R
\name{netEst.dir}
\alias{netEst.dir}
\title{netEST.dir selects the covariance matrix based on external information about 0's and 1's}
\usage{
netEst.dir(X, zero = NULL, one = NULL, lambda, verbose = FALSE, eps = 1e-08)
}
\arguments{
\item{X}{The data matrix of n by p}

\item{zero}{The given external information about 0's}

\item{one}{The given external information about 1's}

\item{lambda}{The tuning parameter in lasso regression, only accept one at a time}

\item{verbose}{Whether to display intermediate outputs.}

\item{eps}{The minimum threshold value for determining edges. By default, it's 1e-8.}
}
\value{
The covariance matrix
}
\description{
netEST.dir selects the covariance matrix based on external information about 0's and 1's
}
\keyword{internal}
