% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgsa_complex.R
\name{netgsa.ftest}
\alias{netgsa.ftest}
\title{netgsa.ftest does this}
\usage{
netgsa.ftest(s2g, s2e, D, DtD, DDInv, n_vec, B, beta_hat)
}
\value{
returns this
}
\description{
more info
}
\keyword{internal}
