% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgsa_complex.R
\name{NetGSA}
\alias{NetGSA}
\title{NetGSA the NetGSA algorithm}
\usage{
NetGSA(
  A,
  x,
  y,
  B,
  lklMethod = c("REML", "ML"),
  directed = FALSE,
  eta = 0.1,
  lim4kappa = 500
)
}
\arguments{
\item{A}{stuff}

\item{x}{stuff}

\item{y}{stuff}

\item{B}{stuff}

\item{directed}{stuff}

\item{eta}{stuff}

\item{lim4kappa}{stuff}

\item{liklMethod}{stuff}
}
\value{
stuff
}
\description{
Runs the NetGSA algorithm for complicated experiments such as:
\enumerate{
\item Estimation for directed and undirected networks
\item Testing for 2 or more conditions
}
}
\keyword{internal}
