% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{netGSAresults}
\alias{netGSAresults}
\alias{netGSAresults<-}
\alias{netGSAresults,DNEAresults-method}
\alias{netGSAresults<-,DNEAresults-method}
\title{Access the netGSA slot of a DNEAresults object}
\usage{
netGSAresults(x)

netGSAresults(x) <- value

\S4method{netGSAresults}{DNEAresults}(x)

\S4method{netGSAresults}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A \code{DNEAresults} object}
}
\value{
A data frame of the results from netGSA.
}
\description{
The function takes as input a \code{DNEAresults} object and returns the netGSA results in the netGSA slot. \cr
}
\examples{
#import example data
data(TEDDYresults)

netGSAresults(TEDDYresults)
}
\seealso{
\code{\link{runNetGSA}}, \code{\link[netgsa:NetGSA]{netgsa::NetGSA()}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
