% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{nodeList}
\alias{nodeList}
\alias{nodeList<-}
\alias{nodeList,DNEAobj-method}
\alias{nodeList<-,DNEAobj-method}
\title{Access the node list}
\usage{
nodeList(x)

nodeList(x) <- value

\S4method{nodeList}{DNEAobj}(x)

\S4method{nodeList}{DNEAobj}(x) <- value
}
\arguments{
\item{x}{a \code{DNEAobj} object}
}
\value{
A data frame corresponding to the node list determined by DNEAdev.
}
\description{
The function takes as input a \code{DNEAobj} object and returns the node list created from \code{\link{createDNEAobject}}. \cr
}
\examples{
#import example data
data(dnw)

nodeList(dnw)
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{clusterNet}}, \code{\link{getNetworkFiles}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
