% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{numFeatures}
\alias{numFeatures}
\title{numFeatures finds the total number of features in the dataset}
\usage{
numFeatures(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
The number of features in the dataset
}
\description{
numFeatures finds the total number of features in the dataset
}
