% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{numSamples}
\alias{numSamples}
\title{numSamples finds the total number of samples in the dataset}
\usage{
numSamples(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
The number of samples in the dataset
}
\description{
numSamples finds the total number of samples in the dataset
}
