% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{optimizedLambda}
\alias{optimizedLambda}
\alias{optimizedLambda<-}
\alias{optimizedLambda,DNEAobj-method}
\alias{optimizedLambda<-,DNEAobj-method}
\title{Access the lambda value used in analysis}
\usage{
optimizedLambda(x)

optimizedLambda(x) <- value

\S4method{optimizedLambda}{DNEAobj}(x)

\S4method{optimizedLambda}{DNEAobj}(x) <- value
}
\arguments{
\item{x}{A \code{DNEAobj} object}

\item{value}{a single-value numeric vector corresponding to the lambda value to use in analysis}
}
\value{
The optimized lambda hyperparameter.
}
\description{
The function takes as input a \code{DNEAobj} object and returns the hyperparameter (lambda) that is currently being
used for the analysis. The user may also provide a single-value numeric vector to change the lambda value for analysis
}
\examples{
#import example data
data(dnw)

optimizedLambda(dnw)
}
\seealso{
\code{\link{BICtune}}
}
\author{
Christopher Patsalis
}
