% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-external.R
\name{plotNetworks}
\alias{plotNetworks}
\title{Visualize the biological networks identified via DNEA}
\usage{
plotNetworks(object, type = c("group_networks", "subnetworks"), subnetwork)
}
\arguments{
\item{object}{A DNEAresults object}

\item{type}{"group_networks" will plot the case, control, and total network. "subnetworks" will plot the subnetwork
specified by teh subnetwork parameter}

\item{subnetwork}{The subnetwork to plot}
}
\value{
a plot of the specified network
}
\description{
The function plots the condition networks as well as the specified subnetworks identified via DNEA
}
\examples{
#import example data
data(TEDDYresults)

#plot the networks
plotNetworks(object = TEDDYresults, type = "group_networks")
plotNetworks(object = TEDDYresults, type = "subnetworks", subnetwork = 1)

}
\seealso{
\code{\link{getNetworks}}, \code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
