% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseNodes-fun.R
\name{reduceFeatures}
\alias{reduceFeatures}
\title{ReduceFeatures}
\usage{
reduceFeatures(
  object,
  method = c("correlation", "knowledge", "hybrid"),
  correlation_threshold = 0.9,
  metabolite_groups = NULL
)
}
\description{
ReduceFeatures
}
