% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDNEAobject.R
\name{restructure_input_data}
\alias{restructure_input_data}
\title{Structure input data for initialization of DNEAobject}
\usage{
restructure_input_data(expression_data, control, case)
}
\arguments{
\item{expression_data}{A matrix or dataframe of un-scaled expression data. The sample names should be rownames
and the feature names should be column names. Column 1 should be a factor of the two conditions, followed by
the numeric expression data}

\item{control}{A string corresponding to the name of condition 1 in column one of the data matrix}

\item{case}{A string corresponding to the name of condition 2 in column one of the data matrix}
}
\value{
A list containing two lists. The first list, named assays, contains the uns-scaled data (if provided)
in position 1 and the scaled data in position 2. The second list, named metadata, contains the metadata parsed
from the input data.
}
\description{
Manipulates input data and creates metadata for initialization of DNEAobject
}
\keyword{internal}
