% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sampleNames.pcorNetwork}
\alias{sampleNames.pcorNetwork}
\title{sampleNames retrieves the sample names from the metadata slot.}
\usage{
sampleNames.pcorNetwork(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
A vector of sample names
}
\description{
This function takes in a DNEAobject and return the sample names located in the
metadata slot of the object.
}
