% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scaledExpressionData}
\alias{scaledExpressionData}
\title{scaledExpressionData retrieves the scaled expression data from the assays slot.}
\usage{
scaledExpressionData(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
the scaled expression matrix
}
\description{
This function takes in a DNEAobject and return the scaled expression matrix located in the
assays slot of the object. You can also use it to input scaled expression data into the DNEAobject
}
