% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{selectionProbabilities}
\alias{selectionProbabilities}
\alias{selectionProbabilities,DNEAobj-method}
\title{Access and set the edge selection probabilities from stabilitySelection()}
\usage{
selectionProbabilities(x)

\S4method{selectionProbabilities}{DNEAobj}(x)
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}
}
\value{
A \code{\link{DNEAobj}} object after filling the
selection_probabilities section of the stable_networks slot.
}
\description{
The function takes as input a \code{\link{DNEAobj}} object and returns an
m x m matrix of selection probabilities for every possible network
edge calculated via \code{\link{stabilitySelection}}.
}
\examples{
#import example data
data(dnw)

selectionProbabilities(dnw)
}
\seealso{
\code{\link{stabilitySelection}},\code{\link{selectionResults}}
}
\author{
Christopher Patsalis
}
