% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{show,DNEAinputSummary-method}
\alias{show,DNEAinputSummary-method}
\title{Display general information about the data present in the DNEAinputSummary slots}
\usage{
\S4method{show}{DNEAinputSummary}(object)
}
\arguments{
\item{object}{A DNEAinputSummary object}
}
\value{
A summary of the input data to \code{\link{createDNEAobject}}.
}
\description{
This function will display the number of samples, number of features, and diagnostics values of the input dataset to a
\code{DNEAresults} object
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
