% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{show,DNEAresults-method}
\alias{show,DNEAresults-method}
\title{Display general information about the data present in the DNEAobject slots}
\usage{
\S4method{show}{DNEAresults}(object)
}
\arguments{
\item{object}{A \code{\link{DNEAresults}} object}
}
\value{
A summary of the information stored in a \code{\link{DNEAresults}} object
}
\description{
This function will display a summary of the information stored within a \code{\link{DNEAresults}} object
}
\examples{
#import example data
data(TEDDYresults)

TEDDYresults

}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
