% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_by_condition}
\alias{split_by_condition}
\title{split_by_condition splits the input data by condition}
\usage{
split_by_condition(dat, condition_levels, condition_by_sample)
}
\arguments{
\item{dat}{a matrix of expression data wherein the samples are rows and features are columns.}

\item{condition_levels}{A list or vector of the unique conditions present in dat}

\item{condition_by_sample}{A list or vector of the condition value corresponding to each
sample.}
}
\value{
A list of expression matrices
}
\description{
split_by_condition will separate the input expression matrix into a list of matrices. Each matrix
corresponds to the expression data for one condition specified by condition_levels
}
\keyword{internal}
