% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{stabilitySelection}
\alias{stabilitySelection}
\title{Performs stability selection to determine probability of feature interactions}
\usage{
stabilitySelection(
  object,
  runParallel = TRUE,
  subSample = FALSE,
  nreps = 50,
  nCores = 4,
  optimal_lambda,
  main.seed = 101
)
}
\arguments{
\item{object}{A DNEA object}

\item{runParallel}{A boolean indicating if stability selection should be run in parallel}

\item{subSample}{A boolean that specifies whether the number of samples are unevenly split
by condition and, therefore, should be adjusted for when randomly sampling.}

\item{nreps}{The total number of reps to perform stability selection. As the sample number
gets smaller, more reps should be run; For datasets under 200 samples, we suggest running
500 reps.}

\item{nCores}{The number of cores available for parallel processing. It is generally optimal
to begin with ~10 reps per core. However, you may need to run one rep per core for optimal
performance with some datasets.}

\item{main.seed}{Sets the seed for random number generation. This ensures reproducibility.}

\item{optimized_lambda}{The optimal lambda value to be used in the model. This parameter is only
necessary if BICtune() is not performed}
}
\value{
DNEAobject containing the stable networks for analysis.
}
\description{
This function randomly samples each condition a number of times specified by nreps. It takes a DNEAobject as input.
The unevenGroups parameter should be set to TRUE if the sample numbers across condition are uneven - this will
ensure that samples are randomly sampled in a way that ensures equal representation in stability selection. This
function can also be run in parallel by specifying the available cores through nCores. The main.seed parameter
is for reproducibility and can be left as default.
}
