% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEM.R
\name{stabsel_init}
\alias{stabsel_init}
\title{stabsel_init initializes static tuning variables to pass to stability selection
(CGM_AHP_stabsel | CGM_AHP_stabsel_subsample)}
\usage{
stabsel_init(listX, nreps)
}
\arguments{
\item{listX}{A list containing matrices of the expression data split by condition. The samples
are in rows and the features are columns.}

\item{nreps}{The number of reps performed in stability selection}
}
\value{
a set of variables corresponding to the following:
number of features (num_features), number of conditions (num_conditions),
a vector containing sample numbers by condition (num_samples),
number of samples in smallest condition (min_num_samples),
a matrix of stability selection results (selection_matrix), and a matrix containing
edge selection (edge_matrix)
an initialized vector for BIC scores (BIC_score), an initialized vector for likelihood (likelihood)
}
\description{
stabsel_init initializes static tuning variables to pass to stability selection
(CGM_AHP_stabsel | CGM_AHP_stabsel_subsample)
}
\keyword{internal}
