% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{summary,consensusClusteringResults-method}
\alias{summary,consensusClusteringResults-method}
\alias{summary.consensusClusteringResults}
\title{Retrieve a summary of a consensusClusteringResults object}
\usage{
\S4method{summary}{consensusClusteringResults}(object)
}
\arguments{
\item{object}{A consensusClusteringResults}
}
\value{
A data frame that corresponds to a summary of the results
of consensus clustering.
}
\description{
The function takes as input a consensusClusteringResults object and
returns a summary of the results of consensus clustering determined
via \code{\link{clusterNet}}.
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
