% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-here.R
\docType{data}
\name{TEDDYresults}
\alias{TEDDYresults}
\title{Example results for DNEA}
\format{
\subsection{\code{TEDDYresults}}{

A DNEA results object after completing a DNEA experiment
}
}
\source{
The data is stored in the \code{\link{DNEA}} package and can be accessed by running data(TEDDYresults) in the console.
}
\usage{
TEDDYresults
}
\description{
This data is an example DNEAresults object after performing DNEA. The experiment was performed using the \code{\link{TEDDY}}
data with 4 reps of stability selection, no subsampling, and default parameters everywhere else. The seed was set to 417.
}
\keyword{datasets}
