% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEMtuning.R
\name{tune_init}
\alias{tune_init}
\title{initialize static tuning variables to pass to CGM_AHP_tune}
\usage{
tune_init(trainX, model, lambda)
}
\arguments{
\item{trainX}{a matrix of expression data wherein the samples are rows and features are columns}

\item{model}{A list of corresponding conditions for the training_data}
}
\value{
a set of variables corresponding to the following variables regarding
the training_data: number of features (p), number of conditions (K), number of lambda
values being tested (N), an a vector containing the number of samples per condition (n),
an initialized vector for BIC scores (BIC_score), an initialized vector for likelihood (likelihood)
}
\description{
initialize static tuning variables to pass to CGM_AHP_tune
}
