% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEM-internals.R
\name{tune_lambda}
\alias{tune_lambda}
\title{Return the results of parameter tuning}
\usage{
tune_lambda(
  lambda_values,
  constant_values,
  FUN,
  trainX,
  testX,
  trainY,
  BIC,
  eps,
  eta,
  BPPARAM = bpparam(),
  BPOPTIONS = bpoptions()
)
}
\arguments{
\item{lambda_values}{A list of values to test while optimizing the lambda parameter}

\item{constant_values}{A list of c constant values to test while optimizing the lambda parameter}

\item{FUN}{A character vector corresponding to the name of the tuning function. the default is
'CGM_AHP_tune'}

\item{trainX}{a matrix of expression data wherein the samples are rows and features are columns.}

\item{testX}{a matrix of expression data wherein the samples are rows and features are columns.
testX should be identical to trainX.}

\item{trainY}{a vector of corresponding conditions for samples in trainX}

\item{BIC}{a boolean indicating whether or not to calculate the BIC score for each lambda.
Default is FALSE}

\item{eps}{A significance cut-off for thresholding network edges
The default value is 1e-06. This value generally should not change}

\item{eta}{default parameter ??. Default is 0.1}

\item{BPPARAM}{A \code{\link{BiocParallel}} object}

\item{BPOPTIONS}{a list of options for BiocParallel created using
the \code{\link[BiocParallel:bpoptions]{bpoptions}} function}
}
\value{
A list containing the following items:
\enumerate{
\item \strong{BIC_guo}: A list of the BIC scores and likelihood value for each lambda tested}
\item \strong{lambda_values}: The lambda values tested
\item \strong{lastar_guo}: The lambda value with the minimum BIC score
\item \strong{ballpark_c}: The value for c constant for the lambda value with the
minimum BIC score
}
\description{
This function returns the results of hyperparameter tuning. Setting "ballpark" to TRUE returns only
the c constant that resulted in the minimum BIC score. Setting "ballpark" to FALSE returns the results
from tuning (the c constants tested, the lambda values derived from the c constants, the BIC scores,
and the likelihood values).
}
\author{
Christopher Patsalis
}
