% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{adjacencyGraph}
\alias{adjacencyGraph}
\alias{adjacencyGraph<-}
\alias{adjacencyGraph,DNEAresults-method}
\alias{adjacencyGraph<-,DNEAresults-method}
\alias{adjacencyGraph,consensusClusteringResults-method}
\alias{adjacencyGraph<-,consensusClusteringResults-method}
\title{Retrieve the adjacency graph for the case, control, or joint network}
\usage{
adjacencyGraph(x, graph)

adjacencyGraph(x, graph) <- value

\S4method{adjacencyGraph}{DNEAresults}(x, graph)

\S4method{adjacencyGraph}{DNEAresults}(x, graph) <- value

\S4method{adjacencyGraph}{consensusClusteringResults}(x, graph)

\S4method{adjacencyGraph}{consensusClusteringResults}(x, graph) <- value
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
An \code{\link{igraph}} graph object corresponding to the specified adjacency graph
}
\description{
The function takes as input a DNEAresults object or consensusClusteringResults object and returns the adjacency graph
made for the case, control, or joint network determined via \code{\link{clusterNet}}. \cr
}
\details{
\emph{FOR INTERNAL USE ONLY -} This function is also used internally to input the adjacency graphs created during
\code{\link{clusterNet}} into the consensusClusteringResults object stored in the consensus_clustering slot of the
DNEAresults object
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
\keyword{internal}
