% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{adjacencyGraph}
\alias{adjacencyGraph}
\alias{adjacencyGraph<-}
\alias{adjacencyGraph,DNEAresults-method}
\alias{adjacencyGraph<-,DNEAresults-method}
\alias{adjacencyGraph,consensusClusteringResults-method}
\alias{adjacencyGraph<-,consensusClusteringResults-method}
\title{Retrieve the adjacency graph for the case, control, or joint network}
\usage{
adjacencyGraph(x, graph)

adjacencyGraph(x, graph) <- value

\S4method{adjacencyGraph}{DNEAresults}(x, graph)

\S4method{adjacencyGraph}{DNEAresults}(x, graph) <- value

\S4method{adjacencyGraph}{consensusClusteringResults}(x, graph)

\S4method{adjacencyGraph}{consensusClusteringResults}(x, graph) <- value
}
\arguments{
\item{x}{A \code{\link{DNEAresults}} or \code{\link{consensusClusteringResults}} object}

\item{graph}{A character string indicating which of the adjacency graphs to return. Values can be "joint_graph" for the
whole graph object, or one of the group values returned by \code{\link{networkGroups}}}
}
\value{
An \code{\link{igraph}} graph object corresponding to the specified adjacency graph
}
\description{
The function  returns the adjacency graph made for the case, control, or joint network determined via
\code{\link{clusterNet}}. \cr
}
\examples{
#import example data
data(TEDDYresults)

adjacencyGraph(TEDDYresults, graph = "DM:case")
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
