% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{adjacencyMatrix}
\alias{adjacencyMatrix}
\alias{adjacencyMatrix<-}
\alias{adjacencyMatrix,DNEAresults-method}
\alias{adjacencyMatrix<-,DNEAresults-method}
\title{Retrieve the weighted or unweighted adjacency matrix}
\usage{
adjacencyMatrix(x, weighted)

adjacencyMatrix(x, weighted) <- value

\S4method{adjacencyMatrix}{DNEAresults}(x, weighted = FALSE)

\S4method{adjacencyMatrix}{DNEAresults}(x, weighted = FALSE) <- value
}
\arguments{
\item{a}{DNEAresults object}
}
\value{
A matrix corresponding to the adjacency matrix specified
}
\description{
The function takes as input a DNEAresults object and returns the weighted or unweighted adjacency matrix
determined via \code{\link{getNetworks}}. \cr
}
\details{
\emph{FOR INTERNAL USE ONLY -} This function is also used internally to input the adjacency_matrices calculated during
\code{\link{getNetworks}} into the adjacency_matrix slot of the DNEAresults object
}
\seealso{
\code{\link{getNetworks}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
