% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{adjacencyMatrix}
\alias{adjacencyMatrix}
\alias{adjacencyMatrix<-}
\alias{adjacencyMatrix,DNEAobj-method}
\alias{adjacencyMatrix<-,DNEAobj-method}
\title{Retrieve the weighted or unweighted adjacency matrix}
\usage{
adjacencyMatrix(x, weighted)

adjacencyMatrix(x, weighted) <- value

\S4method{adjacencyMatrix}{DNEAobj}(x, weighted = FALSE)

\S4method{adjacencyMatrix}{DNEAobj}(x, weighted = FALSE) <- value
}
\arguments{
\item{a}{\code{DNEAobj} object}
}
\value{
A matrix corresponding to the adjacency matrix specified.
}
\description{
The function takes as input a \code{DNEAobj} object and returns the weighted or unweighted adjacency matrix
determined via \code{\link{getNetworks}}. \cr
}
\examples{
#import example data
data(dnw)

adjacencyMatrix(dnw, weighted = TRUE)
}
\seealso{
\code{\link{getNetworks}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
