% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{BICscores}
\alias{BICscores}
\alias{BICscores,pcorNetwork-method}
\title{BICscores returns the BIC scores for each lambda value evaluated}
\usage{
BICscores(x)

\S4method{BICscores}{pcorNetwork}(x)
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
The optimized lambda hyperparameter
}
\description{
The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object and returns
the BIC values for each lambda tested during hyperparameter optimization performed via BICtune().
}
