% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{BICscores}
\alias{BICscores}
\alias{BICscores<-}
\alias{BICscores,DNEAresults-method}
\alias{BICscores<-,DNEAresults-method}
\title{Access the BIC scores for each lambda value evaluated}
\usage{
BICscores(x)

BICscores(x) <- value

\S4method{BICscores}{DNEAresults}(x)

\S4method{BICscores}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
The optimized lambda hyperparameter
}
\description{
The function takes as input a DNEAresults object and returns the BIC values for each lambda tested during
hyperparameter optimization performed via BICtune(). \cr
}
\details{
\emph{FOR INTERNAL USE ONLY -} This function is also used to input BIC scores into the DNEAresults object
within \code{\link{BICtune}}.
}
\seealso{
\code{\link{BICtune}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
