% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{CCsummary}
\alias{CCsummary}
\alias{CCsummary<-}
\alias{CCsummary,DNEAresults-method}
\alias{CCsummary<-,DNEAresults-method}
\title{Retrieves the summary results of consensus clustering}
\usage{
CCsummary(x)

CCsummary(x) <- value

\S4method{CCsummary}{DNEAresults}(x)

\S4method{CCsummary}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A \code{\link{DNEAresults}} object}
}
\value{
A data.frame summary of the consensus clustering results from DNEA
}
\description{
The function takes as input a \code{\link{DNEAresults}} object and returns a summary
of the results of consensus clustering determined via \code{\link{clusterNet}}.
}
\examples{
#import example data
data(TEDDYresults)

CCsummary(TEDDYresults)
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
