% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{clusterNet}
\alias{clusterNet}
\title{Identify metabolic modules within the biological networks using a consensus clustering approach}
\usage{
clusterNet(object, tau = 0.5, max_iterations = 5, eps_threshold = 1e-06)
}
\arguments{
\item{object}{A DNEAresults object}

\item{tau}{The \% agreement threshold among the clustering algorithms for a node to be included in a subnetwork}

\item{max_iterations}{The maximum number of replicates of the clustering algorithms to perform before consensus is reached}

\item{eps_threshold}{A cut-off value thresholding the adjacency matrix. Edges with a partial correlation value below this
threshold will be removed.}
}
\value{
A DNEAobject containing sub-network determinations for the nodes within the input network.
A summary of the consensus clustering results can be viewed using getClusterResults().
Sub-network classification for each node can be found in the node_list slot of the returned
DNEAobject.
}
\description{
This function clusters the adjacency matrix constructed using \code{\link{getNetworks}} via the consensus clustering
approach described in Ma et al. For each iteration, only clusters that meet or exceed the consensus threshold (tau) are recognized. A
new graph is then constructed and a another iteration of clustering is performed until consensus or the maximum number of iterations (maxIter) is reached.
}
\details{
#' Seven clustering algorithms from the \code{\link{igraph}} package are utilized in this consensus clustering approach:
\enumerate{
\item \strong{\code{\link{igraph::cluster_edge_betweenness}}}
\item \strong{\code{\link{igraph::cluster_fast_greedy}}}
\item \strong{\code{\link{igraph::cluster_infomap}}}
\item \strong{\code{\link{igraph::cluster_label_prop}}}
\item \strong{\code{\link{igraph::cluster_louvain}}}
\item \strong{\code{\link{igraph::cluster_walktrap}}}
\item \strong{\code{\link{igraph::cluster_leading_eigen}}}}

These algorithms are performed iteratively on the adjacency matrix constructed using \code{\link{getNetworks}} until a consensus
is reached on resulting subnetwork membership, or the specified max_iterations is reached \emph{(Please see references for more details)}.
}
\examples{
#import example data
data(TEDDY)

#initiate DNEAresults object
DNEA <- createDNEAobject(expression_data = TEDDY,
                         project_name = "TEDDYmetabolomics",
                         case = "DM:case",
                         control = "DM:control")

#optimize lambda parameter
DNEA <- BICtune(object = DNEA, BPPARAM = bpparam())

# perform stability selection
DNEA <- stabilitySelection(object = DNEA, subSample = FALSE, nreps = 500, BPPARAM = bpparam())

#construct the networks
DNEA <- getNetworks(object = DNEA)

#view the edgelist
edgeList(object)

#identify metabolic modules via consensus clustering
DNEA <- clusterNet(object = DNEA)

#we can also plot the subnetworks
plotNetworks(object = DNEA, type = "subnetworks", subnetwork = 1)

}
\references{
Ma J, Karnovsky A, Afshinnia F, Wigginton J, Rader DJ, Natarajan L, Sharma K, Porter AC, Rahman M, He J, Hamm L, Shafi T, Gipson D, Gadegbeku C, Feldman H, Michailidis G, Pennathur S. Differential network enrichment analysis reveals novel lipid pathways in chronic kidney disease. Bioinformatics. 2019 Sep 15;35(18):3441-3452. doi: 10.1093/bioinformatics/btz114. PMID: 30887029; PMCID: PMC6748777. \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6748777/}
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{BICtune}}, \code{\link{stabilitySelection}},
\code{\link{edgeList}}, \code{\link{adjacencyMatrix}}, \code{\link{getNetworks}}, \code{\link{edgeList}},
\code{\link{thresholdNetworks}}, \code{\link{plotNetworks}}
}
\author{
Christopher Patsalis
}
