% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{clusterNet}
\alias{clusterNet}
\title{Identify metabolic modules within the biological networks using a consensus clustering approach}
\usage{
clusterNet(object, tau = 0.5, max_iterations = 5, eps_threshold = 1e-06)
}
\arguments{
\item{object}{A DNEAresults object}

\item{tau}{The \% agreement threshold among the clustering algorithms for a node to be included in a subnetwork}

\item{max_iterations}{The maximum number of replicates of the clustering algorithms to perform before consensus is reached}

\item{eps_threshold}{A cut-off value thresholding the adjacency matrix. Edges with a partial correlation value below this
threshold will be removed.}
}
\value{
A DNEAresults object containing sub-network determinations for the nodes within the input network.
A summary of the consensus clustering results can be viewed using \code{\link{CCsummary}}.
Sub-network membership for each node can be found in the "membership" column of the node list, which can be
viewed using \code{\link{nodeList}}.
}
\description{
This function clusters the jointly estimated adjacency matrix constructed using \code{\link{getNetworks}} via
the consensus clustering approach described in Ma et al (\emph{Please see the \strong{\emph{Details}} section for more
information}) to identify metabolic modules, aka subnetworks, present in the larger networks.
Only subnetworks with consensus that meets or exceeds tau are identified as real.
}
\details{
Seven clustering algorithms from the \code{\link{igraph}} package are utilized in this consensus clustering approach:
\enumerate{
\item \code{\link[igraph:cluster_edge_betweenness]{igraph::cluster_edge_betweenness()}}
\item \code{\link[igraph:cluster_fast_greedy]{igraph::cluster_fast_greedy()}}
\item \code{\link[igraph:cluster_infomap]{igraph::cluster_infomap()}}
\item \code{\link[igraph:cluster_label_prop]{igraph::cluster_label_prop()}}
\item \code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}
\item \code{\link[igraph:cluster_walktrap]{igraph::cluster_walktrap()}}
\item \code{\link[igraph:cluster_leading_eigen]{igraph::cluster_leading_eigen()}}}

For each iteration, node membership in each respective cluster is compared across the algorithms, and only the clusters with \% agreement
greater than tau are kept. A new adjacency graph is then created and clustering is performed again. This occurs iteratively until consensus
on stable subnetworks or the specified max_iterations is reached \emph{(Please see references for more details)}.
}
\examples{
#load BiocParallel
library(BiocParallel)

#import example data
data(TEDDY)

#initiate DNEAresults object
DNEA <- createDNEAobject(expression_data = TEDDY,
                         project_name = "TEDDYmetabolomics",
                         case = "DM:case",
                         control = "DM:control")

#optimize lambda parameter
DNEA <- BICtune(object = DNEA, BPPARAM = bpparam())

# perform stability selection
DNEA <- stabilitySelection(object = DNEA, subSample = FALSE, nreps = 500, BPPARAM = bpparam())

#construct the networks
DNEA <- getNetworks(object = DNEA)

#view the edgelist
edgeList(object)

#identify metabolic modules via consensus clustering
DNEA <- clusterNet(object = DNEA)

#we can also plot the subnetworks
plotNetworks(object = DNEA, type = "subnetworks", subnetwork = 1)

}
\references{
Ma J, Karnovsky A, Afshinnia F, Wigginton J, Rader DJ, Natarajan L, Sharma K, Porter AC, Rahman M, He J, Hamm L, Shafi T, Gipson D, Gadegbeku C, Feldman H, Michailidis G, Pennathur S. Differential network enrichment analysis reveals novel lipid pathways in chronic kidney disease. Bioinformatics. 2019 Sep 15;35(18):3441-3452. doi: 10.1093/bioinformatics/btz114. PMID: 30887029; PMCID: PMC6748777. \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6748777/}
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{BICtune}}, \code{\link{stabilitySelection}},
\code{\link{edgeList}}, \code{\link{adjacencyMatrix}}, \code{\link{getNetworks}}, \code{\link{edgeList}},
\code{\link{filterNetworks}}, \code{\link{plotNetworks}}
}
\author{
Christopher Patsalis
}
