% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{collapsed_DNEAobj-class}
\alias{collapsed_DNEAobj-class}
\alias{collapsed_DNEAobj}
\title{collapsed_DNEAobj}
\value{
A collapsed_DNEAobj object
}
\description{
An s4 class to represent the DNEA workflow, including collapsing
features. This class inherits everything from \code{\link{DNEAobj}}
}
\section{Slots}{

\describe{
\item{\code{original_experiment}}{The DNEAobj object input to
\code{\link{aggregateFeatures}}}

\item{\code{feature_membership}}{A data.frame containing all of the features from
the original input data and their corresponding group membership in the
collapsed data.}
}}

\seealso{
\code{\link{aggregateFeatures}},
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
