% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{datasetSummary}
\alias{datasetSummary}
\alias{datasetSummary<-}
\alias{datasetSummary,DNEAresults-method}
\alias{datasetSummary<-,DNEAresults-method}
\title{Access the dataset_summary slot of a DNEAresults object}
\usage{
datasetSummary(x)

datasetSummary(x) <- value

\S4method{datasetSummary}{DNEAresults}(x)

\S4method{datasetSummary}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEA results object}
}
\value{
The # of samples/features and diagnostic values of the input data calculated by \code{\link{createDNEAobject}}
}
\description{
This function prints to console the number of samples, number of features, and diagnostic values of the input data to
\code{\link{createDNEAobject}} at initiation of the DNEA workflow. \cr
}
\details{
\emph{FOR INTERNAL USE ONLY -} This function is also used to input the dataset_summary slot into the DNEAresults object
within \code{\link{createDNEAobject}}.
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
\keyword{internal}
