% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{datasetSummary}
\alias{datasetSummary}
\alias{datasetSummary<-}
\alias{datasetSummary,DNEAresults-method}
\alias{datasetSummary<-,DNEAresults-method}
\title{datasetSummary is a setter/getter function for the dataset_summary slot of a DNEAresults object}
\usage{
datasetSummary(x)

datasetSummary(x) <- value

\S4method{datasetSummary}{DNEAresults}(x)

\S4method{datasetSummary}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEA results object}
}
\value{
The DNEAinputSummary object if retrieving data, or the input object after populating the dataset_summary
slot
}
\description{
datasetSummary is a setter/getter function for the dataset_summary slot of a DNEAresults object
}
\keyword{internal}
