% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{datasetSummary}
\alias{datasetSummary}
\alias{datasetSummary<-}
\alias{datasetSummary,DNEAresults-method}
\alias{datasetSummary<-,DNEAresults-method}
\title{datasetSummary retrieves the dataset summary}
\usage{
datasetSummary(x)

datasetSummary(x) <- value

\S4method{datasetSummary}{DNEAresults}(x)

\S4method{datasetSummary}{DNEAresults}(x) <- value
}
\arguments{
\item{a}{pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
prints the values to console and returns a list of the aforementioned variables
}
\description{
FOR INTERNAL USE ONLY - The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object
and returns the datasetSummary containing the number of features and samples, and the diagnostic values
}
\keyword{internal}
