% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{datasetSummary,pcorNetwork-method}
\alias{datasetSummary,pcorNetwork-method}
\alias{datasetSummary<-,pcorNetwork-method}
\title{datasetSummary allows you to retrieve the dataset summary or input said summary into the
pcorNetwork, DNEAobject, or DNEAobject_collapsed. For internal use only.}
\usage{
\S4method{datasetSummary}{pcorNetwork}(x)

\S4method{datasetSummary}{pcorNetwork}(x) <- value
}
\arguments{
\item{a}{pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
the input object after filling the dataset_summary slot
}
\description{
datasetSummary allows you to retrieve the dataset summary or input said summary into the
pcorNetwork, DNEAobject, or DNEAobject_collapsed. For internal use only.
}
\keyword{internal}
