% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics<-}
\alias{diagnostics,DNEAresults-method}
\alias{diagnostics<-,DNEAresults-method}
\alias{diagnostics,DNEAinputSummary-method}
\alias{diagnostics<-,DNEAinputSummary-method}
\title{diagnostics retrieves the diagnostic values for the input expression data}
\usage{
diagnostics(x)

diagnostics(x) <- value

\S4method{diagnostics}{DNEAresults}(x)

\S4method{diagnostics}{DNEAresults}(x) <- value

\S4method{diagnostics}{DNEAinputSummary}(x)

\S4method{diagnostics}{DNEAinputSummary}(x) <- value
}
\arguments{
\item{a}{pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
returns the diagnostic values for the input expression data
}
\description{
FOR INTERNAL USE ONLY - The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object
and returns the diagnostic values for the input expression data
}
\keyword{internal}
