% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics<-}
\alias{diagnostics,DNEAresults-method}
\alias{diagnostics<-,DNEAresults-method}
\alias{diagnostics,DNEAinputSummary-method}
\alias{diagnostics<-,DNEAinputSummary-method}
\title{Retrieve the diagnostic values for the input expression data}
\usage{
diagnostics(x)

diagnostics(x) <- value

\S4method{diagnostics}{DNEAresults}(x)

\S4method{diagnostics}{DNEAresults}(x) <- value

\S4method{diagnostics}{DNEAinputSummary}(x)

\S4method{diagnostics}{DNEAinputSummary}(x) <- value
}
\arguments{
\item{a}{DNEAresults object or DNEAinputSummary object}
}
\value{
returns the diagnostic values for the input expression data
}
\description{
This function retrieves teh diagnostic values calculated for the input expression data to \code{\link{createDNEAobject}} \cr
}
\details{
\emph{FOR INTERNAL USE ONLY -} This function is also used to input the diagnostic values into the DNEAresults object
within \code{\link{createDNEAobject}}.
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
\keyword{internal}
