% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-here.R
\docType{data}
\name{dnw}
\alias{dnw}
\title{Example results for DNEAdev}
\format{
A DNEAdev results object after completing a DNEAdev experiment.
}
\source{
The data is stored in the \code{\link{DNEAdev}} package and can be accessed by running data(dnw) in the console.
}
\usage{
data("dnw")
}
\value{
A \code{DNEAresults} object containing the results of a DNEAdev experiment
}
\description{
This data is an example DNEAresults object after performing DNEAdev. The experiment was performed using the \code{\link{TEDDY}}
data with 4 reps of stability selection, no subsampling, and default parameters everywhere else. The seed was set to 417.
}
\keyword{datasets}
