% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{edgeList}
\alias{edgeList}
\alias{edgeList<-}
\alias{edgeList,DNEAobj-method}
\alias{edgeList<-,DNEAobj-method}
\title{Access the edge list}
\usage{
edgeList(x)

edgeList(x) <- value

\S4method{edgeList}{DNEAobj}(x)

\S4method{edgeList}{DNEAobj}(x) <- value
}
\arguments{
\item{x}{a \code{\link{DNEAobj}} object}

\item{value}{a data frame of edges in the network}
}
\value{
A data frame corresponding to the edge list
determined by DNEA
}
\description{
The function takes as input a \code{\link{DNEAobj}} object and returns the
edge list created from \code{\link{getNetworks}}.
}
\examples{
#import example data
data(dnw)

edgeList(dnw)
}
\seealso{
\code{\link{getNetworks}},\code{\link{filterNetworks}},
\code{\link{getNetworkFiles}}
}
\author{
Christopher Patsalis
}
