% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{edgeList}
\alias{edgeList}
\alias{edgeList,pcorNetwork-method}
\title{edgeList returns the edge list created from DNEA}
\usage{
edgeList(x)

\S4method{edgeList}{pcorNetwork}(x)
}
\arguments{
\item{x}{a pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
a data.frame corresponding to the edge list determined by DNEA
}
\description{
The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object and returns
the edge list created from the getNetworks() function
}
