% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{edgeList}
\alias{edgeList}
\alias{edgeList<-}
\alias{edgeList,DNEAresults-method}
\alias{edgeList<-,DNEAresults-method}
\title{Access the edge list}
\usage{
edgeList(x)

edgeList(x) <- value

\S4method{edgeList}{DNEAresults}(x)

\S4method{edgeList}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{a \code{DNEAresults} object}
}
\value{
A data frame corresponding to the edge list determined by DNEAdev.
}
\description{
The function takes as input a \code{DNEAresults} object and returns the edge list created from \code{\link{getNetworks}}. \cr
}
\examples{
#import example data
data(TEDDYresults)

edgeList(TEDDYresults)
}
\seealso{
\code{\link{getNetworks}}, \code{\link{filterNetworks}}, \code{\link{getNetworkFiles}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
