% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{expressionData}
\alias{expressionData}
\alias{expressionData,DNEAresults-method}
\title{Access expression data within a DNEAresults object,}
\usage{
expressionData(x, normalized)

\S4method{expressionData}{DNEAresults}(x, normalized = TRUE)
}
\arguments{
\item{x}{A DNEAresults object}

\item{normalized}{A boolean indicating whether the normalized or original input data should be returned}
}
\value{
The expression matrix specified by the user.
}
\description{
This function accesses the expression data stored in the @assays slot of the DNEAresults object. The output is an
\emph{n x m} matrix with one row for each sample and one column for each feature in the data.
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
