% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{expressionData}
\alias{expressionData}
\alias{expressionData,DNEAresults-method}
\title{Access expression data within a DNEAresults object,}
\usage{
expressionData(x, normalized)

\S4method{expressionData}{DNEAresults}(x, normalized = TRUE)
}
\arguments{
\item{x}{A DNEAresults object}

\item{normalized}{A boolean indicating whether the normalized or original input data should be returned}
}
\value{
The indicated expression expression matrix
}
\description{
The assays slot holds the expression data in a n x m matrix, with one row for each sample, and
one column for each feature.
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
