% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{expressionData}
\alias{expressionData}
\alias{expressionData,pcorNetwork-method}
\title{expressionData is an accessor function for the data in the "assays" slot of a pcorNetwork,
DNEAobject, or DNEAobject_collapsed}
\usage{
expressionData(x, type)

\S4method{expressionData}{pcorNetwork}(x, type = c("input", "normalized"))
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}

\item{type}{"input" will return the original data, and "normalized" will return the scaled data}
}
\value{
The indicated expression expression matrix
}
\description{
The assays slot hold the expression data in a n x m matrix, with one row for each sample, and
one column for each feature.
}
