% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{featureNames}
\alias{featureNames}
\alias{featureNames,DNEAresults-method}
\title{Retrieve the feature names from the metadata slot.}
\usage{
featureNames(x, original = TRUE)

\S4method{featureNames}{DNEAresults}(x, original = TRUE)
}
\arguments{
\item{x}{A DNEAresults object}

\item{original}{"TRUE" returns the original feature names and "FALSE" returns the feature
names that have been modified to avoid errors as a result of special characters.}
}
\value{
A character vector of feature names
}
\description{
This function accesses the feature names stored in the @metadata slot of the DNEAresults object.
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
