% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{featureNames}
\alias{featureNames}
\alias{featureNames,pcorNetwork-method}
\title{featureNames retrieves the feature names from the metadata slot.}
\usage{
featureNames(x, original = TRUE)

\S4method{featureNames}{pcorNetwork}(x, original = TRUE)
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}

\item{original}{"TRUE" returns the original feature names and "FALSE" returns the feature
names that have been modified to avoid errors as a result of special characters.}
}
\value{
A vector of feature names
}
\description{
This function takes in a pcorNetwork, DNEAobject, or DNEAobject_collapsed object
and return the specified feature names located in the metadata slot of the object.
}
