% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getNetworkFiles}
\alias{getNetworkFiles}
\title{getNetworkFiles will save the node and edge information}
\usage{
getNetworkFiles(object, file_path)
}
\arguments{
\item{object}{A DNEAobject}
}
\value{
The same DNEAobject as input and saves the node and edge information as .csv files in
the working directory
}
\description{
This function will save the node and edge information as .csv files in the working directory.
The files are formatted for input into Cytoscape.
}
