% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-external.R
\name{getNetworkFiles}
\alias{getNetworkFiles}
\title{Save network information for input to Cytoscape}
\usage{
getNetworkFiles(object, file_path = NULL)
}
\arguments{
\item{object}{A DNEAresults object}

\item{file_path}{The filepath to save the node and edge lists to. If \strong{NULL}, the files will be saved to the working
directory}
}
\value{
Two .csv files, one for the node list and one for the edge list, saved to the specified file path.
}
\description{
This function will save the node and edge information as .csv files in the working directory.
The files are already formatted for input into Cytoscape.
}
\examples{
#load BiocParallel
library(BiocParallel)

#import example data
data(TEDDY)

#initiate DNEAresults object
DNEA <- createDNEAobject(expression_data = TEDDY,
                         project_name = "TEDDYmetabolomics",
                         case = "DM:case",
                         control = "DM:control")

#optimize lambda parameter
DNEA <- BICtune(object = DNEA, BPPARAM = bpparam())

# perform stability selection
DNEA <- stabilitySelection(object = DNEA, subSample = FALSE, nreps = 5, BPPARAM = bpparam())

#construct the networks
DNEA <- getNetworks(object = DNEA)

#identify metabolic modules via consensus clustering
DNEA <- clusterNet(object = DNEA)

#perform pathway enrichment analysis using netGSA
DNEA <- runNetGSA(object = DNEA)


#save node and edge list for input to cytoscape
getNetworkFiles(DNEA)

}
\seealso{
\code{\link{edgeList}}, \code{\link{nodeList}}
}
\author{
Christopher Patsalis
}
