% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{getNetworks}
\alias{getNetworks}
\title{Construct the GLASSO-based biological Networks}
\usage{
getNetworks(object, optimal_lambda, eps_threshold = 1e-06)
}
\arguments{
\item{object}{A DNEAresults object}

\item{optimal_lambda}{\emph{\strong{OPTIONAL}} The lambda hyperparameter to be used in analysis. Not necessary if \code{\link{BICtune}}
or \code{\link{stabilitySelection}} were already performed}

\item{eps_threshold}{A numeric value between 0 and 1 by which to threshold the partial correlation values for edge identification.
Edges with a partial correlation value below this threshold will be zero'd out from the adjacency matrix.}
}
\value{
A DNEAresults object after populating the adjaceny_matrix and edge_list slots with the corresponding
adjacency_matrix for each sample condition as well as the network edge list.
}
\description{
This function utilizes the lambda parameter tuned using \code{\link{BICtune}} \emph{(may also be user-specified)}
and \emph{optionally} the feature-feature edge selection probabilities using \code{\link{stabilitySelection}} to jointly estimate the biological networks
within the dataset.
}
\examples{
#import example data
data(TEDDY)

#initiate DNEAresults object
DNEA <- createDNEAobject(expression_data = TEDDY,
                         project_name = "TEDDYmetabolomics",
                         case = "DM:case",
                         control = "DM:control")

#optimize lambda parameter
DNEA <- BICtune(object = DNEA, BPPARAM = bpparam())

# perform stability selection
DNEA <- stabilitySelection(object = DNEA, subSample = FALSE, nreps = 5, BPPARAM = bpparam())

#construct the networks
DNEA <- getNetworks(object = DNEA)

#now we can plot the group networks
plotNetworks(object = DNEA, type = "group_networks")

}
\references{
Ma J, Karnovsky A, Afshinnia F, Wigginton J, Rader DJ, Natarajan L, Sharma K, Porter AC, Rahman M, He J, Hamm L, Shafi T, Gipson D, Gadegbeku C, Feldman H, Michailidis G, Pennathur S. Differential network enrichment analysis reveals novel lipid pathways in chronic kidney disease. Bioinformatics. 2019 Sep 15;35(18):3441-3452. doi: 10.1093/bioinformatics/btz114. PMID: 30887029; PMCID: PMC6748777. \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6748777/}

Iyer GR, Wigginton J, Duren W, LaBarre JL, Brandenburg M, Burant C, Michailidis G, Karnovsky A. Application of Differential Network Enrichment Analysis for Deciphering Metabolic Alterations. Metabolites. 2020; 10(12):479. \url{https://doi.org/10.3390/metabo10120479}
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{BICtune}}, \code{\link{stabilitySelection}},
\code{\link{edgeList}}, \code{\link{adjacencyMatrix}}
}
\author{
Christopher Patsalis
}
