% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{getNeworks}
\alias{getNeworks}
\title{Creates the network model using glasso}
\usage{
getNeworks(object, optimal_lambda, eps_threshold = 1e-06)
}
\arguments{
\item{object}{A DNEAobject}

\item{optimal_lambda}{Lambda hyperparameter to be used in analysis. Not necessary if BICtune()
or stabilitySelection() were already run.}

\item{eps_threshold}{A cut-off value thresholding the adjacency matrix. Edges with a partial correlation value below this
threshold will be removed.}
}
\value{
A DNEA object containing an adjacency matrix for the data network,
determined by the glasso model
}
\description{
This function takes in a DNEAobjct and fits a glasso model using the optimized lambda value
determined via BICtune() or otherwise specified. If selection probabilites for each node were
calculated using stabilitySelection(), those are also utilized. An adjacency matrix and a
thresholded adjacency matrix, determiend by the eps value, is output.
}
