% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-external.R
\name{includeMetadata}
\alias{includeMetadata}
\title{Add additional metadata to the DNEAresults object}
\usage{
includeMetadata(object, type = c("sample", "feature"), metadata)
}
\arguments{
\item{object}{A DNEAresults object}

\item{type}{sample or feature metadata}

\item{metadata}{a data.frame containing metadata to add}
}
\value{
A DNEAresults object with the specified additions
}
\description{
This function will take additional metadata and add it to the specified dataframe in the metadata
slot. \strong{\emph{NOTE:}} The rownames of the new metadata must match the order of the input sample names or feature names,
respectively
}
\examples{
#import example data
data(TEDDYresults)

#create sample metadata file
new_metadat <- data.frame(new_group = c(rep("group1", 72),
                          rep("group2", 72)),
                          row.names(sampleNames(TEDDYresults)))

#add new metadata to DNEAresults object
TEDDYresults <- includeMetadata(object = TEDDYresults, type = "sample", metadata = new_metadat)

}
\seealso{
\code{\link{featureNames}}, \code{\link{sampleNames}},
}
\author{
Christopher Patsalis
}
