% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{includeMetadata}
\alias{includeMetadata}
\title{includeMetadata adds info to metadata}
\usage{
includeMetadata(object, type = c("sample", "feature"), metadata)
}
\arguments{
\item{object}{A DNEAobject}

\item{type}{sample or feature metadata}

\item{metadata}{a dataframe containing metadata to add}
}
\value{
A DNEAobject
}
\description{
This function will take additional metadata and add it to the corresponding dataframe in the metadata
slot.
}
