% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{lambdas2Test}
\alias{lambdas2Test}
\alias{lambdas2Test<-}
\alias{lambdas2Test,DNEAresults-method}
\alias{lambdas2Test<-,DNEAresults-method}
\title{Access the lambda values tested during hyperparameter optimization}
\usage{
lambdas2Test(x)

lambdas2Test(x) <- value

\S4method{lambdas2Test}{DNEAresults}(x)

\S4method{lambdas2Test}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A pcorNetwork, DNEAobjct, or DNEAobject_collapsed object}
}
\value{
The lambda values to evaluate in optimization.
}
\description{
The function takes as input a \code{\link{DNEAresults}} object and returns the lambda values that were testing
during hyperparameter optimization performed via \code{\link{BICtune}}. \cr
}
\examples{
#import example data
data(TEDDYresults)

lambdas2Test(TEDDYresults)
}
\seealso{
\code{\link{BICtune}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
