% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSEMtuning.R
\name{matTr}
\alias{matTr}
\title{maTr computes the trace of the matrix}
\usage{
matTr(z)
}
\arguments{
\item{z}{square numerical matrix}
}
\value{
trace of the input matrix (scalar)
}
\description{
maTr computes the trace of the matrix
}
