% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{metaData}
\alias{metaData}
\alias{metaData,DNEAobj-method}
\title{Retrieve metadata stored in DNEAobj}
\usage{
metaData(x, type)

\S4method{metaData}{DNEAobj}(x, type = c("samples", "features"))
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object.}

\item{type}{A character string indicating the type of metadata
to access. Can be "sample" or "feature".}
}
\value{
A data frame of the indicated metadata
}
\description{
This function retrieves the specified metadata data frame stored
in the @metadata slot of the \code{\link{DNEAobj}} object.
}
\examples{
#import example data
data(dnw)

metaData(dnw, type = "sample")
}
\seealso{
\code{\link{includeMetadata}}

\code{\link{createDNEAobject}}, \code{\link{includeMetadata}}

\code{\link{includeMetadata}}
}
\author{
Christopher Patsalis
}
