% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{netGSAresults}
\alias{netGSAresults}
\alias{netGSAresults,DNEAobj-method}
\title{Access the netGSA slot of a DNEAobj object}
\usage{
netGSAresults(x)

\S4method{netGSAresults}{DNEAobj}(x)
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}
}
\value{
A data frame of the results from netGSA.
}
\description{
The function takes as input a \code{\link{DNEAobj}} object and returns
the netGSA results in the netGSA slot.
}
\examples{
#import example data
data(dnw)

netGSAresults(dnw)
}
\seealso{
\code{\link{runNetGSA}},
\code{\link[netgsa:NetGSA]{netgsa::NetGSA()}}
}
\author{
Christopher Patsalis
}
