% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{netGSAresults}
\alias{netGSAresults}
\alias{netGSAresults<-}
\alias{netGSAresults,DNEAresults-method}
\alias{netGSAresults<-,DNEAresults-method}
\title{netGSAresults accesses the netGSA slot of a DNEAresults object}
\usage{
netGSAresults(x)

netGSAresults(x) <- value

\S4method{netGSAresults}{DNEAresults}(x)

\S4method{netGSAresults}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
The results from netGSA
}
\description{
The function takes as input a pcorNetwork, DNEAobject, DNEAobject_collapsed, or consensusClusteringResults object
and returns the netGSA results.
\strong{FOR INTERNAL USE ONLY - The results can also be inserted into the object via the setter functionality of the function.}
}
\keyword{internal}
