% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{networkGroupIDs}
\alias{networkGroupIDs}
\alias{networkGroupIDs,DNEAresults-method}
\title{networkGroupIDs returns the group information for samples in a DNEAobject,
or DNEAobject_collapsed}
\usage{
networkGroupIDs(x)

\S4method{networkGroupIDs}{DNEAresults}(x)
}
\arguments{
\item{x}{A DNEAobject or DNEAobject_collapsed}
}
\value{
A vector of the unique condition labels
}
\description{
This function takes in a DNEAobject or DNEAobject_collapsed and return the unique condition
labels located in the metadata slot of the object. You can also change the sample
variable being used to determine groups.
}
