% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{networkGroupIDs}
\alias{networkGroupIDs}
\alias{networkGroupIDs<-}
\alias{networkGroupIDs,DNEAresults-method}
\title{Access and set the experimental group labels utilized by DNEAdev}
\usage{
networkGroupIDs(x)

networkGroupIDs(x) <- value

\S4method{networkGroupIDs}{DNEAresults}(x)
}
\arguments{
\item{x}{A \code{DNEAresults} object}

\item{value}{a character string name corresponding to a column name of the sample metadata data frame}
}
\value{
A vector of the unique condition labels.
}
\description{
This function accesses the experimental group labels for each sample stored in the @metadata slot of a \code{DNEAresults} object.
}
\examples{
#import example data
data(dnw)

networkGroupIDs(dnw)
}
\seealso{
\code{\link{includeMetadata}}
}
\author{
Christopher Patsalis
}
