% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{networkGroups}
\alias{networkGroups}
\alias{networkGroups<-}
\alias{networkGroups,DNEAresults-method}
\alias{networkGroups<-,DNEAresults-method}
\title{networkGroups retrieves the condition values for each sample from the metadata slot.}
\usage{
networkGroups(x)

networkGroups(x) <- value

\S4method{networkGroups}{DNEAresults}(x)

\S4method{networkGroups}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEAobject, or DNEAobject_collapsed object}
}
\value{
A vector of the condition values
}
\description{
This function takes in a DNEAobject, or DNEAobject_collapsed object and
return the condition values located in the metatdata slot of the object
}
