% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{networkGroups}
\alias{networkGroups}
\alias{networkGroups,DNEAresults-method}
\title{Retrieve the unique group values of the experimental condition}
\usage{
networkGroups(x)

\S4method{networkGroups}{DNEAresults}(x)
}
\arguments{
\item{x}{A DNEAobject, or DNEAobject_collapsed object}
}
\value{
A vector of the condition values.
}
\description{
This function takes in a \code{DNEAresults} object and returns the unique group values of the experimental condition in the dataset
}
\examples{
#import example data
data(TEDDYresults)

networkGroups(TEDDYresults)
}
\seealso{
\code{\link{networkGroupIDs}}, \code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
