% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{nodeList}
\alias{nodeList}
\alias{nodeList<-}
\alias{nodeList,pcorNetwork-method}
\alias{nodeList<-,pcorNetwork-method}
\title{nodeList returns the node list created from DNEA}
\usage{
nodeList(x)

nodeList(x) <- value

\S4method{nodeList}{pcorNetwork}(x)

\S4method{nodeList}{pcorNetwork}(x) <- value
}
\arguments{
\item{x}{a pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
a data.frame corresponding to the node list determined by DNEA
}
\description{
The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object and returns
the node list created from the getNetworks() function
}
\keyword{internal}
