% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{nodeList}
\alias{nodeList}
\alias{nodeList<-}
\alias{nodeList,DNEAresults-method}
\alias{nodeList<-,DNEAresults-method}
\title{Access the node list}
\usage{
nodeList(x)

nodeList(x) <- value

\S4method{nodeList}{DNEAresults}(x)

\S4method{nodeList}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{a \code{\link{DNEAresults}} object}
}
\value{
a data.frame corresponding to the node list determined by DNEA
}
\description{
The function takes as input a \code{\link{DNEAresults}} object and returns the node list created from \code{\link{createDNEAobject}}. \cr
}
\examples{
#import example data
data(TEDDYresults)

nodeList(TEDDYresults)
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{clusterNet}}, \code{\link{getNetworkFiles}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
