% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{numFeatures}
\alias{numFeatures}
\alias{numFeatures,DNEAresults-method}
\alias{numFeatures,DNEAinputSummary-method}
\title{numFeatures finds the total number of features in the dataset}
\usage{
numFeatures(x)

\S4method{numFeatures}{DNEAresults}(x)

\S4method{numFeatures}{DNEAinputSummary}(x)
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
The number of features in the dataset
}
\description{
numFeatures finds the total number of features in the dataset
}
