% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{numFeatures}
\alias{numFeatures}
\alias{numFeatures,DNEAresults-method}
\alias{numFeatures,DNEAinputSummary-method}
\title{Retrieve the total number of features in the dataset}
\usage{
numFeatures(x)

\S4method{numFeatures}{DNEAresults}(x)

\S4method{numFeatures}{DNEAinputSummary}(x)
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
The number of features in the dataset
}
\description{
This function prints to console the total number of features in the dataset
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
