% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{numSamples}
\alias{numSamples}
\alias{numSamples,pcorNetwork-method}
\title{numSamples finds the total number of samples in the dataset}
\usage{
numSamples(x)

\S4method{numSamples}{pcorNetwork}(x)
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
The number of samples in the dataset
}
\description{
numSamples finds the total number of samples in the dataset
}
