% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{numSamples}
\alias{numSamples}
\alias{numSamples,DNEAobj-method}
\alias{numSamples,DNEAinputSummary-method}
\title{Retrieves the total number of samples in the dataset}
\usage{
numSamples(x)

\S4method{numSamples}{DNEAobj}(x)

\S4method{numSamples}{DNEAinputSummary}(x)
}
\arguments{
\item{x}{A \code{DNEAobj} object}
}
\value{
The number of samples in the dataset.
}
\description{
This function prints to console the total number of samples in the dataset
}
\examples{
#import example data
data(dnw)

numSamples(dnw)
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
