% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{optimizedLambda}
\alias{optimizedLambda}
\alias{optimizedLambda<-}
\alias{optimizedLambda,DNEAresults-method}
\alias{optimizedLambda<-,DNEAresults-method}
\title{Access the lambda value used in analysis}
\usage{
optimizedLambda(x)

optimizedLambda(x) <- value

\S4method{optimizedLambda}{DNEAresults}(x)

\S4method{optimizedLambda}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
The optimized lambda hyperparameter
}
\description{
The function takes as input a DNEAresults object and returns the hyperparameter (lambda) that is currently being
used for the analysis. The user may also provide a single-value numeric vector to change the lambda value for analysis
}
\seealso{
\code{\link{BICtune}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
