% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{optimizedLambda}
\alias{optimizedLambda}
\title{optimizedLambda returns the lambda value used in analysis.}
\usage{
optimizedLambda(x)
}
\arguments{
\item{x}{A DNEAobject}
}
\value{
The optimized lambda hyperparameter
}
\description{
optimizedLambda returns the lambda value used in analysis.
}
