% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotNetworks}
\alias{plotNetworks}
\title{plot will create network graphs of the metabolic modules}
\usage{
plotNetworks(object, type = c("group_networks", "subnetworks"), subnetwork)
}
\arguments{
\item{object}{A DNEAresults object}

\item{type}{"group_networks" will plot the case, control, and total network. "subnetworks" will plot the subnetwork
specified by teh subnetwork parameter}

\item{subnetwork}{The subnetwork to plot}
}
\value{
a network plot for each of the subnetworks
}
\description{
The function takes as input a DNEAresults object and creates network plots for each of the subnetworks identified
via runConsensusClustering().
}
