% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotNetworks}
\alias{plotNetworks}
\title{Visualize the biological networks identified via DNEA}
\usage{
plotNetworks(object, type = c("group_networks", "subnetworks"), subnetwork)
}
\arguments{
\item{object}{A DNEAresults object}

\item{type}{"group_networks" will plot the case, control, and total network. "subnetworks" will plot the subnetwork
specified by teh subnetwork parameter}

\item{network}{The subnetwork to plot}
}
\value{
a plot of the specified network
}
\description{
The function plots the condition networks as well as the specified subnetworks identified via DNEA
}
\examples{
#import example data
data(TEDDY)

#initiate DNEAresults object
DNEA <- createDNEAobject(expression_data = TEDDY,
                         project_name = "TEDDYmetabolomics",
                         case = "DM:case",
                         control = "DM:control")

#optimize lambda parameter
DNEA <- BICtune(object = DNEA, BPPARAM = bpparam())

# perform stability selection
DNEA <- stabilitySelection(object = DNEA, subSample = FALSE, nreps = 5, BPPARAM = bpparam())

#construct the networks
DNEA <- getNetworks(object = DNEA)

#identify metabolic modules via consensus clustering
DNEA <- runConsensusCluster(object = DNEA)

#plot the networks
plotNetworks(object = DNEA, type = "group_networks")
plotNetworks(object = DNEA, type = "subnetworks", subnetwork = 1)

}
\seealso{
\code{\link{getNetworks}}, \code{\link{runConsensusCluster}}
}
\author{
Christopher Patsalis
}
