% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotNetworks}
\alias{plotNetworks}
\title{Visualize the biological networks identified via DNEA}
\usage{
plotNetworks(object, type = c("group_networks", "subnetworks"), subnetwork)
}
\arguments{
\item{object}{A DNEAresults object}

\item{type}{"group_networks" will plot the case, control, and total network. "subnetworks" will plot the subnetwork
specified by teh subnetwork parameter}

\item{subnetwork}{The subnetwork to plot}
}
\value{
a plot of the specified network
}
\description{
The function plots the condition networks as well as the specified subnetworks identified via DNEA
}
\seealso{
\code{\link{getNetworks}}, \code{\link{runConsensusCluster}}

Other network construction: 
\code{\link{getNetworks}()}

Other consensus clustering: 
\code{\link{runConsensusCluster}()}
}
\author{
Christopher Patsalis
}
\concept{consensus clustering}
\concept{network construction}
