% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{projectName}
\alias{projectName}
\alias{projectName,DNEAresults-method}
\title{Return the name of the current experiment}
\usage{
projectName(x)

\S4method{projectName}{DNEAresults}(x)
}
\arguments{
\item{x}{A \code{DNEAresults} object}
}
\value{
The name of the DNEAdev experiment.
}
\description{
This function returns the name of the DNEAdev experiment
}
\examples{
#import example data
data(dnw)

projectName(dnw)
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
