% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{projectName}
\alias{projectName}
\alias{projectName<-}
\alias{projectName,DNEAresults-method}
\title{Return the name of the current experiment}
\usage{
projectName(x)

projectName(x) <- value

\S4method{projectName}{DNEAresults}(x)
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
The name of the DNEA experiment
}
\description{
projectName() returns the name of the current experiment and allows the user to change the name as well
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
