% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseNodes-fun.R
\name{ReduceFeatures}
\alias{ReduceFeatures}
\title{ReduceFeatures}
\usage{
ReduceFeatures(
  object,
  method = c("correlation", "knowledge", "hybrid"),
  corr.threshold = 0.9,
  metabolite.groups = NULL
)
}
\description{
ReduceFeatures
}
