% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{runConsensusCluster}
\alias{runConsensusCluster}
\title{runConsensusCluster performs consensus clustering using an adjacency matrix for the network}
\usage{
runConsensusCluster(
  object,
  tau = 0.5,
  num_iterations = 10,
  method = "ensemble"
)
}
\arguments{
\item{object}{A DNEA object}

\item{tau}{The consensus probabilty threshold for agreement among clustering algorithms}

\item{num_iterations}{The number of clustering iterations to perform - this parameter not relevant
for the "ensemble" method. Default is 10 iterations.}

\item{method}{The consensus clustering method to be used. The options are as follows
- "ensemble": indicates that all seven of the available clustering methods
(cluster_edge_betweenness, cluster_fast_greedy, cluster_infomap, cluster_label_prop,
cluster_leading_eigen, cluster_louvain, cluster_walktrap) should be used.
-"lpm" utilizes the cluster_label_prop, cluster_infomap, and cluster_walktrap methods.
-"walktrap" utilizes only cluster_walktrap.
-"infomap" utilizes only infomap}
}
\value{
A DNEAobject containing sub-network determinations for the nodes within the input network.
A summary of the consensus clustering results can be viewed using getClusterResults().
Sub-network classification for each node can be found in the node_list slot of the returned
DNEAobject.
}
\description{
This function will take as input an adjacency matrix from the determined networks and perform
consensus clustering using the following methods from the igraph package: cluster_edge_betweenness,
cluster_fast_greedy, cluster_infomap, cluster_label_prop, cluster_leading_eigen, cluster_louvain,
cluster_walktrap. The output results in sub-network classification for the nodes within the network.
}
