% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{runConsensusCluster}
\alias{runConsensusCluster}
\title{runConsensusCluster performs consensus clustering using an adjacency matrix for the network}
\usage{
runConsensusCluster(
  object,
  tau = 0.5,
  max_iterations = 5,
  eps_threshold = 1e-06
)
}
\arguments{
\item{object}{A DNEA object}

\item{tau}{The consensus probabilty threshold for agreement among clustering algorithms}

\item{max_iterations}{The maximum number of replicates of the clustering algorithms to perform before consensus is reached}

\item{eps_threshold}{A cut-off value thresholding the adjacency matrix. Edges with a partial correlation value below this
threshold will be removed.}
}
\value{
A DNEAobject containing sub-network determinations for the nodes within the input network.
A summary of the consensus clustering results can be viewed using getClusterResults().
Sub-network classification for each node can be found in the node_list slot of the returned
DNEAobject.
}
\description{
This function will take as input an adjacency matrix from the determined networks and perform
consensus clustering using the following methods from the igraph package: cluster_edge_betweenness,
cluster_fast_greedy, cluster_infomap, cluster_label_prop, cluster_leading_eigen, cluster_louvain,
cluster_walktrap. The output results in sub-network classification for the nodes within the network.
}
