% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{runNetGSA}
\alias{runNetGSA}
\title{Runs the NetGSA algorithm described in Hellstern et al.}
\usage{
runNetGSA(object)
}
\arguments{
\item{object}{A DNEAobject}
}
\value{
A DNEAobject containing containing results from NetGSA. Pathway expression differences for
each node can be found in the node_list. A summary of the NetGSA results can be viewed
using getNetGSAresults().
}
\description{
runNetGSA takes as input a DNEA object and utilizes the adjacency matrix calculated in getNetworks()
to perform NetGSA on the two networks to determine differences in subnetwork (pathway) expression.
}
