% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{runNetGSA}
\alias{runNetGSA}
\title{Identify metabolic modules that are enriched across experimental conditions}
\usage{
runNetGSA(object, min_size = 5)
}
\arguments{
\item{object}{A \code{DNEAresults}}

\item{min_size}{The minimum size of metabolic modules for enrichment analysis}
}
\value{
A \code{DNEAresults} object after populating the @netGSA slot. Pathway expression differences for
each node can be found in the node_list. A summary of the NetGSA results can be viewed
using \code{\link{netGSAresults}}.
}
\description{
This function performs pathway enrichment analysis on the metabolic modules identified via \code{\link{clusterNet}}
using the \code{\link[netgsa:NetGSA]{netgsa::NetGSA()}} algorithm.
}
\examples{
#import completed example data
data(TEDDYresults)

#perform pathway enrichment analysis using netGSA
TEDDYresults <- runNetGSA(object = TEDDYresults, min_size = 5)

#view the results
netGSAresults(TEDDYresults)


}
\references{
Hellstern M, Ma J, Yue K, Shojaie A. netgsa: Fast computation and interactive visualization for topology-based pathway enrichment analysis. PLoS Comput Biol. 2021 Jun 11;17(6):e1008979. doi: 10.1371/journal.pcbi.1008979. PMID: 34115744; PMCID: PMC8221786. url{https://pubmed.ncbi.nlm.nih.gov/34115744/}
}
\seealso{
\code{\link{netGSAresults}}, \code{\link{plotNetworks}}, \code{\link{nodeList}}
}
\author{
Christopher Patsalis
}
