% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,DNEAresults-method}
\title{Retrieve the sample names from the metadata slot.}
\usage{
sampleNames(x)

\S4method{sampleNames}{DNEAresults}(x)
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
A vector of sample names
}
\description{
This function takes in a DNEAresults object and return the sample names located in the metadata slot of the object.
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
