% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,pcorNetwork-method}
\title{sampleNames retrieves the sample names from the metadata slot.}
\usage{
sampleNames(x)

\S4method{sampleNames}{pcorNetwork}(x)
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
A vector of sample names
}
\description{
This function takes in a pcorNetwork, DNEAobject, or DNEAobject_collapsed object and
return the sample names located in the metadata slot of the object.
}
