% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,DNEAresults-method}
\title{Retrieve the sample names from the metadata slot.}
\usage{
sampleNames(x)

\S4method{sampleNames}{DNEAresults}(x)
}
\arguments{
\item{x}{A \code{DNEAresults} object}
}
\value{
A character vector of sample names.
}
\description{
This function accesses the sample names stored in the @metadata slot of the \code{DNEAresults} object.
}
\examples{
#import example data
data(TEDDYresults)

sampleNames(TEDDYresults)
}
\seealso{
\code{\link{createDNEAobject}}
}
\author{
Christopher Patsalis
}
