% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{selectionProbabilities}
\alias{selectionProbabilities}
\alias{selectionProbabilities<-}
\alias{selectionProbabilities,DNEAresults-method}
\alias{selectionProbabilities<-,DNEAresults-method}
\title{Access and set the edge selection probabilities from stabilitySelection()}
\usage{
selectionProbabilities(x)

selectionProbabilities(x) <- value

\S4method{selectionProbabilities}{DNEAresults}(x)

\S4method{selectionProbabilities}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
A DNEAresults object after filling the
selection_probabilities section of the stable_networks slot
}
\description{
The function takes as input a DNEAresults object and returns an m x m matrix of selection probabilities for every
possible network edge calculated via \code{\link{stabilitySelection}}. \cr
}
\details{
\emph{FOR INTERNAL USE ONLY -} This function is also used to input the selection probabilities into the DNEAresults object
within \code{\link{stabilitySelection}}.
}
\seealso{
\code{\link{stabilitySelection}}, \code{\link{selectionResults}}
}
\author{
Christopher Patsalis
}
\keyword{helper}
\keyword{internal}
