% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{selectionProbabilities}
\alias{selectionProbabilities}
\alias{selectionProbabilities<-}
\alias{selectionProbabilities,pcorNetwork-method}
\alias{selectionProbabilities<-,pcorNetwork-method}
\title{selectionProbabilities returns an m x m matrix of the edge selection results from stabilitySelection()}
\usage{
selectionProbabilities(x)

selectionProbabilities(x) <- value

\S4method{selectionProbabilities}{pcorNetwork}(x)

\S4method{selectionProbabilities}{pcorNetwork}(x) <- value
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
A pcorNetwork, DNEAobject, or DNEAobject_collapsed object after filling the
selection_probabilities section of the stable_networks slot
}
\description{
The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object and returns
an m x m matrix of selection Probabilities for every possible network edge calculated through
stabilitySelection()
}
\keyword{internal}
