% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{selectionResults}
\alias{selectionResults}
\alias{selectionResults,DNEAobj-method}
\title{Access and set the edge selection results from stabilitySelection()}
\usage{
selectionResults(x)

\S4method{selectionResults}{DNEAobj}(x)
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}
}
\value{
A \code{\link{DNEAobj}} object after filling the
selection_results section of the stable_networks slot.
}
\description{
The function takes as input a \code{\link{DNEAobj}} object and returns an m x m
matrix of selection results for every possible network edge calculated
via \code{\link{stabilitySelection}}.
}
\examples{
#import example data
data(dnw)

selectionResults(dnw)
}
\seealso{
\code{\link{stabilitySelection}},\code{\link{selectionProbabilities}}
}
\author{
Christopher Patsalis
}
