% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{selectionResults}
\alias{selectionResults}
\alias{selectionResults<-}
\alias{selectionResults,pcorNetwork-method}
\alias{selectionResults<-,pcorNetwork-method}
\title{selectionResults returns an m x m matrix of the edge selection results from stabilitySelection()}
\usage{
selectionResults(x)

selectionResults(x) <- value

\S4method{selectionResults}{pcorNetwork}(x)

\S4method{selectionResults}{pcorNetwork}(x) <- value
}
\arguments{
\item{x}{A pcorNetwork, DNEAobject, or DNEAobject_collapsed object}
}
\value{
A pcorNetwork, DNEAobject, or DNEAobject_collapsed object after filling the
selection_results section of the stable_networks slot
}
\description{
The function takes as input a pcorNetwork, DNEAobject, or DNEAobject_collapsed object and returns
an m x m matrix of selection results for every possible network edge calculated through
stabilitySelection()
}
\keyword{internal}
