% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{selectionResults}
\alias{selectionResults}
\alias{selectionResults<-}
\alias{selectionResults,DNEAresults-method}
\alias{selectionResults<-,DNEAresults-method}
\title{Access and set the edge selection results from stabilitySelection()}
\usage{
selectionResults(x)

selectionResults(x) <- value

\S4method{selectionResults}{DNEAresults}(x)

\S4method{selectionResults}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A \code{DNEAresults} object}
}
\value{
A \code{DNEAresults} object after filling the
selection_results section of the stable_networks slot.
}
\description{
The function takes as input a \code{DNEAresults} object and returns an m x m matrix of selection results for every
possible network edge calculated via \code{\link{stabilitySelection}}. \cr
}
\examples{
#import example data
data(TEDDYresults)

selectionResults(TEDDYresults)
}
\seealso{
\code{\link{stabilitySelection}}, \code{\link{selectionProbabilities}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
