% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{show,DNEAresults-method}
\alias{show,DNEAresults-method}
\title{Display general information about the data present in the DNEAobject slots}
\usage{
\S4method{show}{DNEAresults}(object)
}
\arguments{
\item{object}{A DNEAobject}
}
\value{
A summary of the information stored in a DNEAresults object
}
\description{
This function will display a summary of the information stored within a DNEAresults object
}
\seealso{
\code{\link{createDNEAobject}}, \code{\link{reduceFeatures}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
