% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{subnetworkMembership}
\alias{subnetworkMembership}
\alias{subnetworkMembership,DNEAobj-method}
\alias{subnetworkMembership,consensusClusteringResults-method}
\title{Retrieve the subnetwork membership for each feature}
\usage{
subnetworkMembership(x)

\S4method{subnetworkMembership}{DNEAobj}(x)

\S4method{subnetworkMembership}{consensusClusteringResults}(x)
}
\arguments{
\item{x}{A \code{\link{DNEAobj}} object}
}
\value{
A data frame that corresponds to the results of
consensus clustering.
}
\description{
The function takes as input a \code{\link{DNEAobj}} or
\code{consensusClusteringResults} object and returns the results of
consensus clustering determined via \code{\link{clusterNet}}.
}
\examples{
#import example data
data(dnw)

subnetworkMembership(dnw)
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
