% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{subnetworkMembership}
\alias{subnetworkMembership}
\alias{subnetworkMembership,consensusClusteringResults-method}
\alias{subnetworkMembership,DNEAresults-method}
\title{Retrieve the subnetwork membership for each feature}
\usage{
subnetworkMembership(x)

\S4method{subnetworkMembership}{consensusClusteringResults}(x)

\S4method{subnetworkMembership}{DNEAresults}(x)
}
\arguments{
\item{x}{A DNEAresults object}
}
\value{
A data.frame that corresponds to the results of consensus clustering
}
\description{
The function takes as input a DNEAresults or consensusClusteringResults object
and returns the results of consensus clustering determined via \code{\link{clusterNet}}.
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
