% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/all-methods.R
\name{subnetworkMembership}
\alias{subnetworkMembership}
\alias{subnetworkMembership<-}
\alias{subnetworkMembership,consensusClusteringResults-method}
\alias{subnetworkMembership,DNEAresults-method}
\alias{subnetworkMembership<-,DNEAresults-method}
\title{Retrieve the subnetwork membership for each feature}
\usage{
subnetworkMembership(x)

subnetworkMembership(x) <- value

\S4method{subnetworkMembership}{consensusClusteringResults}(x)

\S4method{subnetworkMembership}{DNEAresults}(x)

\S4method{subnetworkMembership}{DNEAresults}(x) <- value
}
\arguments{
\item{x}{A \code{DNEAresults} object}
}
\value{
A data frame that corresponds to the results of consensus clustering.
}
\description{
The function takes as input a \code{DNEAresults} or \code{consensusClusteringResults} object
and returns the results of consensus clustering determined via \code{\link{clusterNet}}.
}
\examples{
#import example data
data(TEDDYresults)

subnetworkMembership(TEDDYresults)
}
\seealso{
\code{\link{clusterNet}}
}
\author{
Christopher Patsalis
}
\keyword{internal}
