% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-here.R
\docType{data}
\name{TEDDY}
\alias{TEDDY}
\title{The example data from xxx}
\format{
\subsection{\code{TEDDY} A data frame with 100 rows and 144 columns. Each row corresponds to a sample, and each column}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{    corresponds to a metabolite.
}\if{html}{\out{</div>}}
}
}
\source{
The raw data can be downloaded from the Metabolomics workbench under study ID \strong{ST001386}:
\url{https://www.metabolomicsworkbench.org/data/DRCCStudySummary.php?Mode=SetupRawDataDownload&StudyID=ST001386}
}
\usage{
data("TEDDY")
}
\value{
An \emph{m x n} expression matrix of metabolomics data from the TEDDY dataset
}
\description{
This data is an \emph{m x n} expression matrix corresponding to a curated list of metabolites from "The Environmental Determinants of
Diabetes in the Young" clinical trial.
}
\references{
Lee HS, Burkhardt BR, McLeod W, Smith S, Eberhard C, Lynch K, Hadley D, Rewers M, Simell O, She JX, Hagopian B, Lernmark A, Akolkar B, Ziegler AG, Krischer JP; TEDDY study group. Biomarker discovery study design for type 1 diabetes in The Environmental Determinants of Diabetes in the Young (TEDDY) study. Diabetes Metab Res Rev. 2014 Jul;30(5):424-34. doi: 10.1002/dmrr.2510. PMID: 24339168; PMCID: PMC4058423. \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4058423/}
}
\keyword{datasets}
